### 2019-03-18 #1329 @whedon
[PRE REVIEW]: corporaexplorer: an R package for dynamic exploration of text collections
**Submitting author:** @kgjerde (<a href="http://orcid.org/0000-0001-8291-6837">Kristian Lundby Gjerde</a>)
**Repository:** <a href="https://github.com/kgjerde/corporaexplorer" target ="_blank">https://github.com/kgjerde/corporaexplorer</a>
**Version:** v0.4.0
