### 2019-03-19 f770ba1
Fixed function name in documentation
