+++ R/prepare_data.R
@@ -277,7 +277,7 @@ get_term_vector <- function(returned_list) {
 #' corpus <- prepare_data(test_df, corpus_name = "Test corpus")
 #' \dontrun{
 #' # Running exploration app:
-#' run_corpus_exploration(corpus)
+#' run_corpus_explorer(corpus)
 #'
 #' # Running app to extract documents:
 #' run_document_extractor(corpus)
