+++ README.Rmd
@@ -55,6 +55,15 @@ devtools::install_github("kgjerde/corporaexplorer")
 ```
 
 
+## Example corpora
+
+The package includes a tiny 'corporaexplorerobject' (see below): `corporaexplorer::test_data`, which consists of 10 tiny documents and can be used for testing the basic functionality of the package.
+
+A more typical corpus can be downloaded for testing [here](https://ndownloader.figshare.com/files/14626970)
+-- this is a 'corporaexplorerobject' based on a document collection scraped from the Russian governmental website kremlin.ru (more info [here](https://figshare.com/s/820c105c12755d853801)).
+The equivalent material in Russian can be downloaded [here](https://figshare.com/s/be0d2cfa195e43160fcb).<sup>[2](#myfootnote2)</sup>
+
+
 ## Functions
 
 **corporaexplorer** contains three functions:
@@ -65,13 +74,6 @@ exploration of a 'corporaexplorerobject'.
 3. `run_document_extractor()` runs a Shiny app for simple retrieval/extraction of
 documents from a 'corporaexplorerobject' in a reading-friendly format.
 
-## Example corpora
-
-The package includes a tiny 'corporaexplorerobject': `corporaexplorer::test_data`, which consists of 10 tiny documents and can be used for testing the basic functionality of the package.
-
-A more typical corpus can be downloaded for testing [here](https://ndownloader.figshare.com/files/14626970)
--- this is a 'corporaexplorerobject' based on a document collection scraped from the Russian governmental website kremlin.ru (more info [here](https://figshare.com/s/820c105c12755d853801)).
-The equivalent material in Russian can be downloaded [here](https://figshare.com/s/be0d2cfa195e43160fcb).<sup>[2](#myfootnote2)</sup>
 
 
 ## 1. Prepare data for the Shiny apps
@@ -220,7 +222,7 @@ knitr::include_graphics("man/figures/wall.png")
 
 ### 2d. Advanced detail: Regular expression engines
 
-`run_corpus_exploration()` lets you choose among three regex engine setups:
+`run_corpus_explorer()` lets you choose among three regex engine setups:
 
 1. `default`: use the [`re2r`](https://github.com/qinwf/re2r) package for simple searches and the [`stringr`](https://github.com/tidyverse/stringr) package for complex regexes (details below).
 This is the recommended option.
@@ -237,7 +239,7 @@ Please consult the documentation for [`re2r`](https://github.com/qinwf/re2r) and
 
 By default, searches for patterns consisting of a single word and without special characters will be carried out in a document term matrix. Other searches are carried out in full text.
 
-Advanced users can set the `optional_info` parameter in `run_corpus_exploration()` to `TRUE`:
+Advanced users can set the `optional_info` parameter in `run_corpus_explorer()` to `TRUE`:
 this will print to console the following information for each input term:
 which regex engine was used,
 and whether the search was carried out in the document term matrix or in the full text documents.
