+++ man/run_corpus_explorer.Rd
@@ -54,7 +54,7 @@ test_df <- tibble::tibble(Date = dates, Text = texts, Title = titles)
 corpus <- prepare_data(test_df, corpus_name = "Test corpus")
 \dontrun{
 # Running exploration app:
-run_corpus_exploration(corpus)
+run_corpus_explorer(corpus)
 
 # Running app to extract documents:
 run_document_extractor(corpus)
