+++ README.Rmd
@@ -60,8 +60,8 @@ devtools::install_github("kgjerde/corporaexplorer")
 The package includes a tiny 'corporaexplorerobject' (see below): `corporaexplorer::test_data`, which consists of 10 tiny documents and can be used for testing the basic functionality of the package.
 
 A more typical corpus can be downloaded for testing [here](https://ndownloader.figshare.com/files/14626970)
--- this is a 'corporaexplorerobject' based on a document collection scraped from the Russian governmental website kremlin.ru (more info [here](https://figshare.com/s/820c105c12755d853801)).
-The equivalent material in Russian can be downloaded [here](https://figshare.com/s/be0d2cfa195e43160fcb).<sup>[2](#myfootnote2)</sup>
+-- this is a 'corporaexplorerobject' based on a document collection scraped from the Russian governmental website kremlin.ru (7751 documents, more info [here](https://figshare.com/s/820c105c12755d853801)).
+The equivalent material in Russian (10221 documents) can be downloaded [here](https://figshare.com/s/be0d2cfa195e43160fcb).<sup>[2](#myfootnote2)</sup>
 
 
 ## Functions
