### 2019-03-21 #1329 @trinker
I'd be happy to review as well.

On Thu, Mar 21, 2019, 6:59 PM Kenneth Benoit <notifications@github.com>
wrote:

> Looks very cool, I'd be happy to contribute a review.
>
