### 2019-03-21 fb2c95a
Added info about example corpora [skip ci]
