+++ README.Rmd
@@ -61,7 +61,8 @@ The package includes a tiny 'corporaexplorerobject' (see below): `corporaexplore
 
 A more typical corpus can be downloaded for testing [here](https://ndownloader.figshare.com/files/14626970)
 -- this is a 'corporaexplorerobject' based on a document collection scraped from the Russian governmental website kremlin.ru (7751 documents, more info [here](https://figshare.com/s/820c105c12755d853801)).
-The equivalent material in Russian (10221 documents) can be downloaded [here](https://figshare.com/s/be0d2cfa195e43160fcb).<sup>[2](#myfootnote2)</sup>
+The equivalent material in Russian (10,221 documents) can be downloaded [here](https://figshare.com/s/be0d2cfa195e43160fcb).
+A third, somewhat bigger (but less carefully prepared) test corpus (26,654 documents) from the same website is available [here](https://figshare.com/s/fcd93f6f93bb23be3bb1).<sup>[2](#myfootnote2)</sup>
 
 
 ## Functions
