### 2019-03-21 #1329 @kgjerde
@leouieda, @kbenoit, @trinker: Thanks a lot to all of you for such quick responses.

Also: I just added a link to one more example corpus in the readme file, but will leave the Github repo alone now.
