### 2019-03-21 #1342 @whedon
[REVIEW]: corporaexplorer: an R package for dynamic exploration of text collections
**Submitting author:** @kgjerde (<a href="http://orcid.org/0000-0001-8291-6837">Kristian Lundby Gjerde</a>)
**Repository:** <a href="https://github.com/kgjerde/corporaexplorer" target ="_blank">https://github.com/kgjerde/corporaexplorer</a>
**Version:** 0.5.1
