### 2019-03-22 #1329 @leouieda
:wave: @kbenoit @trinker thank you for agreeing to review :tada: Please head over to #1342 for the review. There are checklists of criteria for both of you. You can also leave comments on the issue with your review or ask any questions. Also feel free to open issues and make pull requests on the software repository. If you do so, please link back to the review issue by mentioning `openjournals/joss-reviews#1342`.

@kgjerde thank you for recommending reviewers :+1: Please do your best to respond to the reviewers comments and suggestions. And don't hesitate to ask if you have any questions or concerns.
