### 2019-03-30 kgjerde/corporaexplorer#1 @trinker
Improve statement of need
@kgjerde the review for this work asks:

> Do the authors clearly state what problems the software is designed to solve and who the target audience is?

in regard to the documentation.  Looking at the readme you have clearly stated the audience.  Could you make the problem it solves more explicit?  I surmise that this work enables the scaling up of analysis to larger corpora.  I think the argument could be made that the package also shines a light on the places where a qualitative researcher might look closer.  it enables the researcher to traverse more ground.

In any event you have named the target audience, can you write a few lines in the README to convince them (A) they have a problem and that (B) this tool solves that problem.
