### 2019-03-30 42c6df5
Improve statement of need in readme.
Addresses #1
