### 2019-03-30 kgjerde/corporaexplorer#1 @kgjerde
@trinker Thank you for looking at this! I have addressed the issue in 42c6df5. Please let me know if you think this needs further improvement.
