### 2019-03-31 kgjerde/corporaexplorer#2 @trinker
Possible Permissions problem with README link
Under the *Example corpora* heading in the README I see:

> A more typical corpus can be downloaded for testing here –

with the text hyperlink: https://ndownloader.figshare.com/files/14626970.  This results in:

```
{"message": "Insufficient permissions", "code": "InsufficientPermissions"}
```
