### 2019-03-31 kgjerde/corporaexplorer#3 @trinker
Possible encoding issue
I ran:

```
library(corporaexplorer)

dates <- as.Date(paste(2011:2020, 1:10, 21:30, sep = "-"))
texts <- paste0(
  "This is a document about ", month.name[1:10], ". ",
  "This is not a document about ", rev(month.name[1:10]), "."
)
titles <- paste("Text", 1:10)
test_df <- tibble::tibble(Date = dates, Text = texts, Title = titles)

# Converting to corporaexplorer object:
corpus <- prepare_data(test_df, corpus_name = "foo")
```

Inside of the shiny app I see the `Corpus map â€“ foo`.  I assume there's some sort of encoding issue here.  Note I am on a Windows 10 machine.

![image](https://user-images.githubusercontent.com/1763278/55283135-16eeee80-532a-11e9-9b4e-d6131b09254a.png)
