### 2019-03-31 kgjerde/corporaexplorer#4 @trinker
`prepare_data` function creates tables with column headers of different casing
The `prepare_data` function creates tables with column headers of different casing.  For example
I run this minimal example:

```
library(corporaexplorer)

dates <- as.Date(paste(2011:2020, 1:10, 21:30, sep = "-"))
texts <- paste0(
  "This is a document about ", month.name[1:10], ". ",
  "This is not a document about ", rev(month.name[1:10]), "."
)
titles <- paste("Text", 1:10)
test_df <- tibble::tibble(Date = dates, Text = texts, Title = titles)

# Converting to corporaexplorer object:
corpus <- prepare_data(test_df, corpus_name = "foo")

lapply(corpus[[1]], names)
```

And see:

```
$data_dok
[1] "Date"      "Text"      "Title"     "Year"      "wc"        "id"        "Text_case"

$data_365
 [1] "Date"      "Year"      "empty"     "wc"        "Weekday_n" "Month_n"   "Week_n"    "Yearday_n" "No."       "ID"        "mnd_vert"  "mnd_hor"   "id"
```

There's upper, lower, and snake all mixed together.  It's nice to fix these inconsistencies (unless there's a reason for them) right from the start.
