### 2019-03-31 kgjerde/corporaexplorer#3 @trinker
Similar in `Corpus Info` tab following `Search results chart`.

![image](https://user-images.githubusercontent.com/1763278/55283360-46542a00-532f-11e9-90f2-137746ed79c3.png)
