### 2019-03-31 kgjerde/corporaexplorer#3 @trinker
And again in the document pane after the word `Document`

![image](https://user-images.githubusercontent.com/1763278/55283398-4143aa80-5330-11e9-8c80-0a69be0fa469.png)
