### 2019-03-31 kgjerde/corporaexplorer#5 @trinker
Clarity on labels and descriptions around search terms effect of chart filling and document highlighting.
The term *highlight* gets used in different ways that was confusing to me at first.  In the documentation of the README I see:

> Term(s) to chart: The corpus map plot will highlight days/documents where these terms are present. The terms will also be highlighted in all documents.
> Add terms for text highlighting: Terms included here (one on each line) will be highlighted in all documents. An arbitrary number of terms can be added.

And in the app I see:

![image](https://user-images.githubusercontent.com/1763278/55283421-d21a8600-5330-11e9-845e-539b46c5f45c.png)

I realize that `Term(s) to chart` might better be labeled as `Term(s) to chart & highlight` and `Add terms for text highlighting` might better be labeled `Additional terms for text highlighting`.  This makes it clear in the app that the first widget will both fill in the calander chart and highlight the terms in the document viewer.  I also think calling *color fill* in the README *highlighting* is confusing.  I'd recommend refering to filling as one temr and highlighting as another.
