### 2019-03-31 kgjerde/corporaexplorer#6 @trinker
Consider adding documentation around contributing 
As part of the JOSS review process one of the areas to consider is:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I do not see this in the README.  If I missed it my apologies.  If it isn't there consider adding it.  Here is an example of this info in the README from quanteda: https://github.com/quanteda/quanteda#contributing
