### 2019-03-31 kgjerde/corporaexplorer#8 @trinker
4 choices for "Number of terms to chart" is misleading
The documentation for the app clearly states that 2 is the maximum number of terms to chart.

> Number of terms to chart: How many terms (phrases/patterns) should be charted in the corpus map? Current maximum is two.

But the app itself suggest I can chart 4 terms which leads to an error message.

![image](https://user-images.githubusercontent.com/1763278/55283590-adc0a880-5334-11e9-8f2b-c17d488c56f0.png)

Might it be better to  not even allow the user to see 3 or 4 as an option?  Documentation is great but when the app itself constrains behavior by design it makes the interaction nicer.
