### 2019-04-08 aeffa58
Add info about contribution.
- Addresses #6
- Taken from quanteda
