### 2019-04-08 334f68f
BUGFIX. Fixes bug caused by b36b838
- The bug caused month shifts in calendar plot to be faulty, causing error in chronoly in first week of month
