### 2019-04-08 4a6966b
Distinguish between plot 'highlighting' and text highlighting in README [skip ci]
- Addresses #5
- Now uses 'colour fill(ing)' about the plots and 'highlight' about text highlighting
