### 2019-04-08 kgjerde/corporaexplorer#5 @kgjerde
@trinker Thanks for this useful input. I have tried to distinguish more clearly between the two in 4a6966b.
