### 2019-04-08 e219ef2
Clarify "highlighting" in sidebar
Addresses #5
