### 2019-04-08 161fe60
Adjust 'number of terms to chart' option in sidebar.
Adresses #8
[skip ci]
