+++ inst/explorer/ui/ui_sidebar.R
@@ -2,8 +2,9 @@ shinydashboard::dashboardSidebar(
     shinyWidgets::radioGroupButtons(
         inputId = "antall_linjer",
         label = "No. of terms to chart",
-        choices = 1:4,
+        choices = 1:2,
         justified = TRUE,
+        width = '160px',
         size = "xs"
     ),
 
