### 2019-04-08 kgjerde/corporaexplorer#8 @kgjerde
Yes, you are of course right, @trinker. For now, I just adjusted the UI options (in 161fe60), as suggested.
