### 2019-04-08 eb25855
Wording in readme re #6 [skip ci]
