### 2019-04-08 kgjerde/corporaexplorer#6 @kgjerde
Implemented in commits referred to above.
