### 2019-04-08 7bbbc74
Fix link in readme
- Adresses #2
