+++ README.Rmd
@@ -66,7 +66,8 @@ devtools::install_github("kgjerde/corporaexplorer")
 
 The package includes a tiny 'corporaexplorerobject' (see below): `corporaexplorer::test_data`, which consists of 10 tiny documents and can be used for testing the basic functionality of the package.
 
-A more typical corpus can be downloaded for testing [here](https://ndownloader.figshare.com/files/14626970)
+A more typical corpus can be downloaded for testing
+[here](https://ndownloader.figshare.com/files/14626970?private_link=820c105c12755d853801)
 -- this is a 'corporaexplorerobject' based on a document collection scraped from the Russian governmental website kremlin.ru (7751 documents, more info [here](https://figshare.com/s/820c105c12755d853801)).
 The equivalent material in Russian (10,221 documents) can be downloaded [here](https://figshare.com/s/be0d2cfa195e43160fcb).
 A third, somewhat bigger (but less carefully prepared) test corpus (26,654 documents) from the same website is available [here](https://figshare.com/s/fcd93f6f93bb23be3bb1).<sup>[2](#myfootnote2)</sup>
