### 2019-04-08 kgjerde/corporaexplorer#2 @kgjerde
Sorry about that. Fixed.
