+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: corporaexplorer
 Type: Package
 Title: A Shiny App for Exploration of Text Collections
-Version: 0.4.0
+Version: 0.4.0.9000
 Authors@R: person("Kristian Lundby", "Gjerde", email = "kristian.gjerde@gmail.com",
   role = c("aut", "cre"))
 Description: An R package for dynamic exploration of text collections. The
