### 2019-04-08 #1342 @leouieda
:wave: Hi @kbenoit @trinker, just checking in on the progress of this review.

@trinker thank you for posting the issues on the repo :1st_place_medal:  I see that you edited the initial review comment to link to the issues. Please post these as a comment on this issue instead. Otherwise I don't get notifications and it's difficult to keep track of the review progress.
