### 2019-04-08 #1342 @trinker
> 👋 Hi @kbenoit @trinker, just checking in on the progress of this review.
>
> @trinker thank you for posting the issues on the repo 🥇 I see that you edited the initial review comment to link to the issues. Please post these as a comment on this issue instead. Otherwise I don't get notifications and it's difficult to keep track of the review progress.

Will do.  Makes sense on the notifications.  This might be better clarified in the review process as it says:

> Reviewers and authors:
> Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the target repository and link to those issues (especially acceptance-blockers) in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

Maybe jsut add the phrase "as a reply" in the thread below.
