### 2019-04-11 #1342 @trinker
All Elements of the Review Checklist I had raised have been addressed.

  - [x] https://github.com/kgjerde/corporaexplorer/issues/1
  - [x] https://github.com/kgjerde/corporaexplorer/issues/2
  - [x] https://github.com/kgjerde/corporaexplorer/issues/5
  - [x] https://github.com/kgjerde/corporaexplorer/issues/8
  - [x] https://github.com/kgjerde/corporaexplorer/issues/6

### Other

  - [ ] App Encoding issue: https://github.com/kgjerde/corporaexplorer/issues/3
  - [ ] `prepare_date()` object case inconsistencies: https://github.com/kgjerde/corporaexplorer/issues/4
