### 2019-04-11 #1342 @leouieda
@trinker thank you! (Also, I followed up with your suggestion and it has already been included for future JOSS submissions :100:)

@kgjerde please try to address or respond to the last two remaining issues.
