### 2019-04-11 be32c83
Restructuring corporaexploreronject
Addresses #4

data_dok:
* renamed wc --> Tile_length
* renamed id --> ID
* renamed Text_case --> Text_original_case
* changed order of columns, so columns always present go first

data_365:
* renamed wc --> Tile_length
* renamed Text_case --> Text_original_case
* renamed empty --> Day_without_docs
* renamed ID --> Invisible_fake_date and changed from integer to
    boolean (== 0 to TRUE, != 0 to FALSE)
* renamed id --> ID
* Variables no longer exported from function:
    * No., Week_n, mnd_vert, mnd_hor, Month_n, Yearday_n
