+++ R/prepare_data.R
@@ -37,15 +37,30 @@ transform_regular <- function(df, columns_to_include = NULL, normalise = TRUE) {
 
   new_df$Year <- lubridate::year(new_df$Date)
 
-  new_df$wc <- nchar(new_df$Text) %>%
     scales::rescale(to = c(1, 10))
+  new_df$Tile_length <- nchar(new_df$Text) %>%
 
-  new_df$id <- seq_len(nrow(new_df))
+  new_df$ID <- seq_len(nrow(new_df))
 
-  new_df$Text_case <- new_df$Text
+  new_df$Text_original_case <- new_df$Text
 
   new_df$Text <-
     stringr::str_to_lower(new_df$Text) # for søke-purposes
+
+  new_df <- dplyr::select(new_df,
+                ID,
+                Date,
+                Text,
+                Text_original_case,
+                Tile_length,
+                Year,
+                sort(colnames(new_df)[!colnames(new_df) %in% c("ID",
+                                                               "Date",
+                                                               "Text",
+                                                               "Text_original_case",
+                                                               "Tile_length",
+                                                               "Year")]))
+
   message("1/3 Document data frame done.")
   return(new_df)
 }
@@ -53,17 +68,6 @@ transform_regular <- function(df, columns_to_include = NULL, normalise = TRUE) {
 
 
 # 2. 365: 1 day = 1 tile --------------------------------------------------
-
-#' Check if date is the last in the month
-#'
-#' @param date Date.
-#'
-#' @return Logical.
-#' @keywords internal
-is_last_in_month <- function(date) {
-  check <- as.Date(date) + 1
-  lubridate::month(check) != lubridate::month(date)
-}
 
 #' Convert "data_dok" tibble to "data_365" tibble
 #'
@@ -77,7 +81,7 @@ transform_365 <- function(new_df) {
 
   df_365 <- new_df %>%
     dplyr::group_by(Date) %>%
-    dplyr::summarise(Text = paste(Text_case, collapse = "\n\n--\n\n"))
+    dplyr::summarise(Text = paste(Text_original_case, collapse = "\n\n--\n\n"))
 
   min_date <- min(df_365$Date)
   max_date <- max(df_365$Date)
@@ -91,13 +95,13 @@ transform_365 <- function(new_df) {
       end_val = as.Date(paste0(max(df_365$Year), "-12-31"))
     )
 
-  df_365$empty <- is.na(df_365$Text)
-  # df_365$empty[df_365$empty == TRUE] <- "black"
-  df_365$empty[df_365$empty == FALSE] <- NA
+  df_365$Day_without_docs <- is.na(df_365$Text)
+  # df_365$Day_without_docs[df_365$Day_without_docs == TRUE] <- "black"
+  df_365$Day_without_docs[df_365$Day_without_docs == FALSE] <- NA
 
   df_365$Text <- NULL
 
-  df_365$wc <- 1
+  df_365$Tile_length <- 1
 
   df_365$Year <- lubridate::year(df_365$Date)
 
@@ -105,22 +109,14 @@ transform_365 <- function(new_df) {
 
   df_365$Month_n <- lubridate::month(df_365$Date)
 
-  df_365$Week_n <- lubridate::isoweek(df_365$Date)
-
   df_365$Yearday_n <- lubridate::yday(df_365$Date)
-
-  df_365$No. <- 1:nrow(df_365)
 
   df_365 <-
     dplyr::arrange(df_365, Year, Weekday_n, Yearday_n, Month_n)
-  df_365$ID <- 1:nrow(df_365)
 
-  df_365$mnd_vert <-
[…]
