+++ inst/download/global/config.R
@@ -6,7 +6,7 @@ abc <- eval(as.name(shiny::getShinyOption("corporaexplorer_download_data")))
 if (class(abc) == "corporaexplorerobject") {
   INFO_COLUMNS <- abc$columns_for_info
   abc <- abc$original_data$data_dok
-  abc$Text <- abc$Text_case
+  abc$Text <- abc$Text_original_case
 
 }
 
