+++ inst/explorer/server/3_event_corpus_map_click/display_document_text.R
@@ -1,4 +1,4 @@
 output$doc_tekst <- shiny::renderText({
-    display_document(session_variables[[plot_mode$mode]]$Text_case[min_rad],
+    display_document(session_variables[[plot_mode$mode]]$Text_original_case[min_rad],
                      search_arguments)
 })
