### 2019-04-11 5cfeb53
Ensure backwards compatibility
Because be32c83c introduced breaking changes in the corporaexplorerobject, this commit converts "old" objects on the fly on app start-up

Ref #4
