### 2019-04-11 53b3746
Create print method for corporaexplorerobject
Ref #10
