+++ man/print.corporaexplorerobject.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/prepare_data.R
+\name{print.corporaexplorerobject}
+\alias{print.corporaexplorerobject}
+\title{Print corporaexplorerobject}
+\usage{
+\method{print}{corporaexplorerobject}(obj)
+}
+\arguments{
+\item{obj}{A corporaexplorerobject}
+}
+\value{
+Console-friendly output
+}
+\description{
+Print corporaexplorerobject
+}
+\keyword{internal}
