### 2019-04-11 kgjerde/corporaexplorer#4 @kgjerde
I used this suggestion to do a housecleaning of variables that no longer need to be part of the corporaexplorerobject (but were necessary in earlier iterations) and have generally improved naming in be32c83. I then addressed backwards compatibility problems in 5cfeb53.
