### 2019-04-12 40b4eae
Fix "–" encoding problem on windows
Addresses #3
