+++ inst/explorer/server/3_event_corpus_map_click/UI_element_control_data_dok.R
@@ -5,9 +5,9 @@ if (session_variables$doc_list_open == TRUE) {
     shiny::removeTab('dokumentboks',
               target = "document_list_title",
               session = shiny::getDefaultReactiveDomain())
-    
-    
-    
+
+
+
     session_variables$doc_list_open <- FALSE
 }
 
@@ -22,42 +22,42 @@ if (session_variables$doc_tab_open == FALSE) {
                 shiny::plotOutput("dok_vis",
                            click = "dok_vis_click",
                            height = "auto"),
-                
+
                 shiny::htmlOutput(
                     outputId = "doc_tekst",
                     container = shiny::tags$div,
                     class = "boxed_doc"
                 )
             ),
-        
+
         select = TRUE,
         menuName = NULL,
         session = shiny::getDefaultReactiveDomain()
     )
-    
+
     appendTab(
         'dokumentboks',
         tab =
             shiny::tabPanel(
                 title = shiny::textOutput('document_box_title_info'),
                 value = "document_box_title_info",
-                
+
                 shiny::htmlOutput(
                     outputId = "doc_info",
                     container = shiny::tags$div,
                     class = "boxed_doc"
                 )
-                
+
             )
     )
-    
+
     # Setter fane-tittel
     output$document_box_title_info <- shiny::renderText({
         paste0("Document information")
     })
-    
+
     session_variables$doc_tab_open <- TRUE
-    
+
 }
 
 show_ui("document_box")
@@ -65,8 +65,7 @@ show_ui("document_box")
 # Setter boks-tittel
 output$document_box_title <- shiny::renderText({
     paste0("Document \u2013 ",
-           as.character(format(
-               session_variables[[plot_mode$mode]]$Date[min_rad],
-               "%A %d %B %Y"
-           )))
+           format_date(
+               session_variables[[plot_mode$mode]]$Date[min_rad]),
+           )
 })
