### 2019-04-14 748f961
Add cautionary note about character encoding, in particular Windows.
