### 2019-04-14 #1342 @kgjerde
@trinker I think I have addressed all the points you raised – thanks a lot for the very useful input, much appreciated!

In addition, your point on Windows encoding issues (kgjerde/corporaexplorer#3) forced me to set up a Windows environment, which turned out to be very useful. (I had relied on AppVeyor CI to verify the basic functionality on Windows, having only access to Mac OS and Linux myself.) Playing around a bit on Windows, I realized that unicode datasets that work fine on Mac and Linux may play less nicely with Windows, due to encoding issues. I therefore decided to add a note about this in the README (in https://github.com/kgjerde/corporaexplorer/commit/748f961). Any feedback appreciated.
