### 2019-04-14 #1342 @trinker
@kgjerde I tested and they address the issues I raised.
