### 2019-04-14 e6e2540
Corpus info tab now understands that corpus is not subset when 'subset filter' field is hidden but contains terms.
