### 2019-04-14 fe91a5a
Equivalent to 5cfeb53, just for document_extractor (compatibility
