### 2019-04-14 06a4f69
extractor ui sidebar change
- to avoid confusion as to whether subset is 'on' or 'off'
