### 2019-04-14 #1342 @kgjerde
@trinker Yes, the UI issues you thankfully pointed out was quite easy to fix as soon as I got aware of the problem! Which led me to understand that similar issues may arise for the corpus texts as well, if one is unlucky/not careful with encoding – hence the README note.
