### 2019-04-15 7d4ea12
Remove redundant code
