### 2019-04-16 ff614eb
Simplify transform_365
