### 2019-04-19 76e4fdf
readme stylistic changes
