+++ README.Rmd
@@ -8,9 +8,9 @@ output:
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>",
-  fig.path = "man/figures/README-",
-  out.width = "100%"
+  comment = '#>',
+  fig.path = 'man/figures/README-',
+  out.width = '100%'
 )
 ```
 # corporaexplorer
@@ -57,14 +57,14 @@ or who knows anyone with this knowledge.
 To install **corporaexplorer**, run the following from an R console:
 
 ``` r
-install.packages("devtools")
-devtools::install_github("kgjerde/corporaexplorer")
+install.packages('devtools')
+devtools::install_github('kgjerde/corporaexplorer')
 ```
 
 
 ## A note on platforms and encoding
 
-**corporaexplorer** works on Mac OS, Windows and Linux,<sup>[1](#myfootnote1)</sup>
+**corporaexplorer** works on Mac OS, Windows and Linux,<sup>[1](#footnote1)</sup>
 and there are some important differences in how R handles text
 on the different platforms.
 If you are working with plain English text,
@@ -75,7 +75,7 @@ non-[ASCII](https://en.wikipedia.org/wiki/ASCII) encoded text in R
 
 **On Mac OS or Linux**, problems with encoding will likely not arise at all.
 If problems do arise, they can typically be solved by
-making the R 'locale' unicode-friendly (e.g. `Sys.setlocale("LC_ALL", "en_US.UTF-8")`).
+making the R 'locale' unicode-friendly (e.g. `Sys.setlocale('LC_ALL', 'en_US.UTF-8')`).
 NB! This assumes that the text is UTF-8 encoded,
 so if changing the locale in this way does not help,
 make sure that the text is encoded as UTF-8 characters.
@@ -102,7 +102,7 @@ A more typical corpus can be downloaded for testing
 [here](https://ndownloader.figshare.com/files/14626970?private_link=820c105c12755d853801)
 -- this is a 'corporaexplorerobject' based on a document collection scraped from the Russian governmental website kremlin.ru (7751 documents, more info [here](https://figshare.com/s/820c105c12755d853801)).
 The equivalent material in Russian (10,221 documents) can be downloaded [here](https://figshare.com/s/be0d2cfa195e43160fcb).
-A third, somewhat bigger (but less carefully prepared) test corpus (26,654 documents) from the same website is available [here](https://figshare.com/s/fcd93f6f93bb23be3bb1).<sup>[2](#myfootnote2)</sup>
+A third, somewhat bigger (but less carefully prepared) test corpus (26,654 documents) from the same website is available [here](https://figshare.com/s/fcd93f6f93bb23be3bb1).<sup>[2](#footnote2)</sup>
 
 
 ## Functions
@@ -146,17 +146,17 @@ that are relevant for advanced users.
 
 ### 2a. Sidebar input
 
-<img align="left" src="man/figures/sidebar.png" width="170" height="450" style="margin-right: 25px"/>
+<img align='left' src='man/figures/sidebar.png' width='170' height='450' style='margin-right: 25px'/>
 
 * **Number of terms to chart**: How many terms (phrases/patterns) should be charted in the corpus map? Current maximum is two.
 * **Term(s) to chart**: Using colour fillings, the corpus map plot will indicate days/documents where these terms are present. The terms will also be highlighted in all documents.
 * **Additional terms for text highlighting**: Terms included here (one on each line) will be highlighted in all documents. An arbitrary number of terms can be added.
[…]
