+++ README.md
@@ -61,14 +61,14 @@ a rudimentary knowledge of R – or who knows anyone with this knowledge.
 To install **corporaexplorer**, run the following from an R console:
 
 ``` r
-install.packages("devtools")
-devtools::install_github("kgjerde/corporaexplorer")
+install.packages('devtools')
+devtools::install_github('kgjerde/corporaexplorer')
 ```
 
 ## A note on platforms and encoding
 
 **corporaexplorer** works on Mac OS, Windows and
-Linux,<sup>[1](#myfootnote1)</sup> and there are some important
+Linux,<sup>[1](#footnote1)</sup> and there are some important
 differences in how R handles text on the different platforms. If you are
 working with plain English text, there will most likely be no issues
 with encoding on any platform. Unfortunately, working with
@@ -78,8 +78,8 @@ Windows.
 
 **On Mac OS or Linux**, problems with encoding will likely not arise at
 all. If problems do arise, they can typically be solved by making the R
-‘locale’ unicode-friendly (e.g. `Sys.setlocale("LC_ALL",
-"en_US.UTF-8")`). NB\! This assumes that the text is UTF-8 encoded, so
+‘locale’ unicode-friendly (e.g. `Sys.setlocale('LC_ALL',
+'en_US.UTF-8')`). NB\! This assumes that the text is UTF-8 encoded, so
 if changing the locale in this way does not help, make sure that the
 text is encoded as UTF-8 characters. Alternatively, if you can ascertain
 the character encoding, set the locale correspondingly.
@@ -110,7 +110,7 @@ material in Russian (10,221 documents) can be downloaded
 [here](https://figshare.com/s/be0d2cfa195e43160fcb). A third, somewhat
 bigger (but less carefully prepared) test corpus (26,654 documents) from
 the same website is available
-[here](https://figshare.com/s/fcd93f6f93bb23be3bb1).<sup>[2](#myfootnote2)</sup>
+[here](https://figshare.com/s/fcd93f6f93bb23be3bb1).<sup>[2](#footnote2)</sup>
 
 ## Functions
 
@@ -160,7 +160,7 @@ users.
 
 ### 2a. Sidebar input
 
-<img align="left" src="man/figures/sidebar.png" width="170" height="450" style="margin-right: 25px"/>
+<img align='left' src='man/figures/sidebar.png' width='170' height='450' style='margin-right: 25px'/>
 
   - **Number of terms to chart**: How many terms (phrases/patterns)
     should be charted in the corpus map? Current maximum is two.
@@ -173,19 +173,19 @@ users.
   - **Filter corpus?** Input terms here (one on each line) will filter
     the corpus so that it includes only documents where all these terms
     are present. Unlike terms entered in the two fields above, these
-    terms will not be included in the summary statistics in the “Corpus
-    info” tab. An arbitrary number of terms can be used.
+    terms will not be included in the summary statistics in the ‘Corpus
+    info’ tab. An arbitrary number of terms can be used.
   - **Case sensitive search**: Check this box to distinguish between
     lower and upper case in the search. Unchecked, `war` will find both
-    “war” and “War”; if checked, `war` will only find “war”.
-  - **“Year range” or “date range”**: Filters the corpus (the first and
+    ‘war’ and ‘War’; if checked, `war` will only find ‘war’.
+  - **‘Year range’ or ‘date range’**: Filters the corpus (the first and
     last date included) by time span.
-  - **Plot mode (“calendar” or “document wall”)**: Should the resulting
+  - **Plot mode (‘calendar’ or ‘document wall’)**: Should the resulting
     corpus map treat one day or one document as its base unit?
   - **Adjust plot size**: Change plot height. This is the only sidebar
     input that has effect without clicking the search button.
[…]
