+++ README.md
@@ -27,7 +27,7 @@ package makes it easier to interact with larger text collections than
 what is usually feasible, and it facilitates creativity by allowing for
 more open-ended explorations of the text collections. Further, it can
 make selection and extraction of texts for further scrutiny faster than
-a ‘traditional workflow’ allows, and it facilitates transparency and
+a “traditional workflow” allows, and it facilitates transparency and
 rigidity by making it easier to account for or reproduce the process of
 selecting texts for close reading. Above all, the package is aimed at
 flexibility: it does not enforce any given workflow, but can play a
@@ -61,8 +61,8 @@ a rudimentary knowledge of R – or who knows anyone with this knowledge.
 To install **corporaexplorer**, run the following from an R console:
 
 ``` r
-install.packages('devtools')
-devtools::install_github('kgjerde/corporaexplorer')
+install.packages("devtools")
+devtools::install_github("kgjerde/corporaexplorer")
 ```
 
 ## A note on platforms and encoding
@@ -78,8 +78,8 @@ Windows.
 
 **On Mac OS or Linux**, problems with encoding will likely not arise at
 all. If problems do arise, they can typically be solved by making the R
-‘locale’ unicode-friendly (e.g. `Sys.setlocale('LC_ALL',
-'en_US.UTF-8')`). NB\! This assumes that the text is UTF-8 encoded, so
+“locale” unicode-friendly (e.g. `Sys.setlocale("LC_ALL",
+"en_US.UTF-8")`). NB\! This assumes that the text is UTF-8 encoded, so
 if changing the locale in this way does not help, make sure that the
 text is encoded as UTF-8 characters. Alternatively, if you can ascertain
 the character encoding, set the locale correspondingly.
@@ -88,21 +88,21 @@ the character encoding, set the locale correspondingly.
 thing is to check carefully that the texts appear as expected in
 `corporaexplorer`’s apps, and that the searches function as expected. If
 there are problems, a good place to start is a blog post with the
-telling title [‘Escaping from character encoding hell in R on
-Windows’](https://dss.iq.harvard.edu/blog/escaping-character-encoding-hell-r-windows).
+telling title [“Escaping from character encoding hell in R on
+Windows”](https://dss.iq.harvard.edu/blog/escaping-character-encoding-hell-r-windows).
 
 For (a lot) more information about encoding, see [this informative
 article](http://kunststube.net/encoding/) by David C. Zentgraf.
 
 ## Example corpora
 
-The package includes a tiny ‘corporaexplorerobject’ (see below):
+The package includes a tiny “corporaexplorerobject” (see below):
 `corporaexplorer::test_data`, which consists of 10 tiny documents and
 can be used for testing the basic functionality of the package.
 
 A more typical corpus can be downloaded for testing
 [here](https://ndownloader.figshare.com/files/14626970?private_link=820c105c12755d853801)
-– this is a ‘corporaexplorerobject’ based on a document collection
+– this is a “corporaexplorerobject” based on a document collection
 scraped from the Russian governmental website kremlin.ru (7751
 documents, more info
 [here](https://figshare.com/s/820c105c12755d853801)). The equivalent
@@ -116,11 +116,11 @@ the same website is available
 
 **corporaexplorer** contains three functions:
 
-1.  `prepare_data()` converts a data frame to a ‘corporaexplorerobject’.
+1.  `prepare_data()` converts a data frame to a “corporaexplorerobject”.
 2.  `run_corpus_explorer()` runs the package’s core feature, a Shiny app
-    for fast and flexible exploration of a ‘corporaexplorerobject’.
+    for fast and flexible exploration of a “corporaexplorerobject”.
 3.  `run_document_extractor()` runs a Shiny app for simple
-    retrieval/extraction of documents from a ‘corporaexplorerobject’ in
[…]
