+++ README.Rmd
@@ -8,9 +8,9 @@ output:
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = '#>',
-  fig.path = 'man/figures/README-',
-  out.width = '100%'
+  comment = "#>",
+  fig.path = "man/figures/README-",
+  out.width = "100%"
 )
 ```
 # corporaexplorer
@@ -31,7 +31,7 @@ rely on close reading of textual documents as part of their academic activity, a
 **corporaexplorer** aims to enhance such workflows in several ways.
 The package makes it easier to interact with larger text collections than what is usually feasible,
 and it facilitates creativity by allowing for more open-ended explorations of the text collections.
-Further, it can make selection and extraction of texts for further scrutiny faster than a 'traditional workflow' allows,
+Further, it can make selection and extraction of texts for further scrutiny faster than a "traditional workflow" allows,
 and it facilitates transparency and rigidity by making it easier to account for or reproduce the process of selecting texts for close reading.
 Above all, the package is aimed at flexibility: it does not enforce any given workflow, but can play a small or larger role in many different research designs.
 
@@ -57,8 +57,8 @@ or who knows anyone with this knowledge.
 To install **corporaexplorer**, run the following from an R console:
 
 ``` r
-install.packages('devtools')
-devtools::install_github('kgjerde/corporaexplorer')
+install.packages("devtools")
+devtools::install_github("kgjerde/corporaexplorer")
 ```
 
 
@@ -75,7 +75,7 @@ non-[ASCII](https://en.wikipedia.org/wiki/ASCII) encoded text in R
 
 **On Mac OS or Linux**, problems with encoding will likely not arise at all.
 If problems do arise, they can typically be solved by
-making the R 'locale' unicode-friendly (e.g. `Sys.setlocale('LC_ALL', 'en_US.UTF-8')`).
+making the R "locale" unicode-friendly (e.g. `Sys.setlocale("LC_ALL", "en_US.UTF-8")`).
 NB! This assumes that the text is UTF-8 encoded,
 so if changing the locale in this way does not help,
 make sure that the text is encoded as UTF-8 characters.
@@ -87,7 +87,7 @@ the texts appear as expected in `corporaexplorer`'s apps,
 and that the searches function as expected.
 If there are problems, a good place to start is a
 blog post with the telling title
-['Escaping from character encoding hell in R on Windows'](https://dss.iq.harvard.edu/blog/escaping-character-encoding-hell-r-windows).
+["Escaping from character encoding hell in R on Windows"](https://dss.iq.harvard.edu/blog/escaping-character-encoding-hell-r-windows).
 
 For (a lot) more information about encoding, see
 [this informative article](http://kunststube.net/encoding/)
@@ -96,11 +96,11 @@ by David C. Zentgraf.
 
 ## Example corpora
 
-The package includes a tiny 'corporaexplorerobject' (see below): `corporaexplorer::test_data`, which consists of 10 tiny documents and can be used for testing the basic functionality of the package.
+The package includes a tiny "corporaexplorerobject" (see below): `corporaexplorer::test_data`, which consists of 10 tiny documents and can be used for testing the basic functionality of the package.
 
 A more typical corpus can be downloaded for testing
 [here](https://ndownloader.figshare.com/files/14626970?private_link=820c105c12755d853801)
--- this is a 'corporaexplorerobject' based on a document collection scraped from the Russian governmental website kremlin.ru (7751 documents, more info [here](https://figshare.com/s/820c105c12755d853801)).
[…]
