### 2019-04-21 7190c18
Implement possibility to edit legend keys for corpus info plot
Re #7
