+++ inst/explorer/server/2_event_search_button/UI_element_control.R
@@ -18,6 +18,12 @@ output$dag_kart <-
 hide_ui("day_corpus_box")
 hide_ui("document_box")
 
+# Edit info plot legend keys UI
+shinyWidgets::updateCheckboxGroupButtons(session,
+                                         "edit_info_plot_legend_keys",
+                                         selected = "No")
+shinyjs::hideElement("edit_info_plot_legend_keys")
+
 # Update date inputs, making sure they remain within corpus date range
 if (search_arguments$time_filtering_mode == "Year range") {
   updateDateRangeInput(
