+++ inst/explorer/server/session_variables.R
@@ -8,7 +8,8 @@ session_variables <- shiny::reactiveValues(
     doc_list_open = FALSE,
     created_info = FALSE,
     stop_info_tab = TRUE,
-    day_plot_height = EMPTY_DAY_PLOT_HEIGHT
+    day_plot_height = EMPTY_DAY_PLOT_HEIGHT,
+    corpus_info_plot = NULL
 )
 
 search_arguments <- shiny::reactiveValues(
