+++ inst/explorer/server/corpus_info_tab.R
@@ -3,12 +3,12 @@
 shiny::observe({
   if (input$corpus_box == "Corpus info" &
     session_variables$created_info == FALSE) {
-
     if (search_arguments$all_ok) {
-
-      start_df_list <- create_df_for_info(session_variables,
-                                          search_arguments,
-                                          plot_mode$mode)
+      start_df_list <- create_df_for_info(
+        session_variables,
+        search_arguments,
+        plot_mode$mode
+      )
     }
 
     # Info til infofane
@@ -28,9 +28,13 @@ shiny::observe({
       )
     })
 
+# Table title -------------------------------------------------------------
+
     output$search_results <- shiny::renderText({
       if (session_variables$stop_info_tab == FALSE) {
         if (highlight_terms_exist(search_arguments)) {
+          shinyjs::showElement("edit_info_plot_legend_keys")
+
           if (corpus_is_filtered(
             search_arguments,
             session_variables,
@@ -50,6 +54,8 @@ shiny::observe({
       }
     })
 
+# Table -------------------------------------------------------------------
+
     output$TABLE <- shiny::renderTable({
       if (session_variables$stop_info_tab == FALSE) {
         if (highlight_terms_exist(search_arguments)) {
@@ -61,6 +67,8 @@ shiny::observe({
         }
       }
     }, na = "\u2013", digits = 2, width = "100%")
+
+# Plot title --------------------------------------------------------------
 
     output$info_plot_title <- shiny::renderText({
       if (session_variables$stop_info_tab == FALSE) {
@@ -85,15 +93,81 @@ shiny::observe({
     })
 
 
+# Plot --------------------------------------------------------------------
+
     output$corpus_info_plot <- shiny::renderPlot({
       if (session_variables$stop_info_tab == FALSE) {
-      if (nrow(start_df_list$start_df) != 0)  {
-        corpus_info_plot(start_df_list, search_arguments)
-      }
+        if (nrow(start_df_list$start_df) != 0) {
+          session_variables$corpus_info_plot <- corpus_info_plot(start_df_list, search_arguments)
+
+          session_variables$corpus_info_plot
+        }
       } else {
         NULL
       }
     })
+
+# Edit legend key text input UI -------------------------------------------
+
+    output$column_info_names_ui <- shiny::renderUI({
+      ui_list <- lapply(seq_along(start_df_list$full_terms), function(i) {
+        id <- paste0("info_legend_", i)
+        shiny::textInput(id,
+          label = NULL,
+          value = start_df_list$full_terms[i]
+        )
+      })
+
+      for (i in seq_along(start_df_list$full_terms)) {
+        ui_list[[i]][["children"]][[2]][["attribs"]][["style"]] <-
+          paste0(
+            "
+  padding-left:10px;
+  padding-top:10px;
+  padding-right:10px;
+  padding-bottom:10px;
+  margin-top:-1em;
+  margin-bottom:-1em;
+  resize:none;
+  height:25px;
+  color:white;
+  width:75%;
+  ",
+            paste0("background-color:", MY_COLOURS[i])
+          )
+      }
+
+      ui_list
+    })
+
+
     session_variables$created_info <- TRUE
   }
+})
+
+# Event: update button ----------------------------------------------------
+
+shiny::observeEvent(input$update_info_plot_legend_keys, {
+  new_legend_terms <- collect_info_plot_legends(search_arguments$terms_highlight)
+
+  output$corpus_info_plot <- shiny::renderPlot({
+    shiny::validate(shiny::need(
+      length(new_legend_terms) == length(unique(new_legend_terms)),
[…]
