### 2019-04-24 8309d71
Freshened up update_session_variables
- thus also avoiding error that would occur if search/highlight terms include duplicates + manual editing of info plot legend keys (due to len(unique terms) != len(all terms)
