+++ inst/explorer/server/2_event_search_button/update_session_variables.R
@@ -1,5 +1,4 @@
 # "Reset" corpus before each search ---------------------------------------
-
 session_variables$data_365 <-
     loaded_data$original_data$data_365
 session_variables$data_dok <-
@@ -16,10 +15,13 @@ session_variables$stop_info_tab <- TRUE
 # And corpus info plot
 session_variables$corpus_info_plot <- NULL
 
+
 # Collect search arguments from sidebar -----------------------------------
 
+# Calendar or document wall -----------------------------------------------
 plot_mode$mode <- input$modus
 
+# Time --------------------------------------------------------------------
 search_arguments$time_filtering_mode <- input$years_or_dates
 
 if (search_arguments$time_filtering_mode == "Year range") {
@@ -28,23 +30,45 @@ if (search_arguments$time_filtering_mode == "Year range") {
     search_arguments$time_range <- c(input$date_calendar[1], input$date_calendar[2])
 }
 
+# Case sensitive? ---------------------------------------------------------
 search_arguments$case_sensitive <- input$case_sensitivity
 
+# Filtering corpus --------------------------------------------------------
 if (!is.null(input$subset_corpus)) {
     search_arguments$subset_terms <- collect_subset_terms()
-    search_arguments$subset_tresholds <-collect_treshold_values(search_arguments$subset_terms)
+    search_arguments$subset_tresholds <- collect_treshold_values(search_arguments$subset_terms)
     search_arguments$subset_custom_column <- collect_custom_column(search_arguments$subset_terms)
-    search_arguments$subset_terms <- clean_terms(search_arguments$subset_terms)
+    search_arguments$subset_terms <-
+        to_lower_if_case_insensitive_search(search_arguments$subset_terms) %>%
+        clean_terms()
 }
 
-search_arguments$search_terms <- collect_search_terms()
-search_arguments$search_terms <- clean_terms(search_arguments$search_terms)
+# Search terms ------------------------------------------------------------
+search_arguments$search_terms <- collect_search_terms() %>%
+    to_lower_if_case_insensitive_search() %>%
+    unique() %>%
+    clean_terms()
 
-search_arguments$terms_highlight <- collect_highlight_terms()
+#  Highlight terms - collect ----------------------------------------------
+raw_highlight_terms <- collect_highlight_terms()
 
-search_arguments$tresholds <- collect_treshold_values(search_arguments$terms_highlight)
-search_arguments$custom_column <- collect_custom_column(search_arguments$terms_highlight)
+# Highlight terms - step 1 -------------------------------------------------
+search_arguments$terms_highlight <- raw_highlight_terms %>%
+    to_lower_if_case_insensitive_search()
 
-search_arguments$terms_highlight <- clean_terms(search_arguments$terms_highlight)
+# Identifying duplicates --------------------------------------------------
+highlight_terms_not_duplicated <- !duplicated(search_arguments$terms_highlight)
 
+# Highlight terms - step 2 ------------------------------------------------
+search_arguments$terms_highlight <-
+    search_arguments$terms_highlight[highlight_terms_not_duplicated] %>%
+    clean_terms()
+
+# Search arguments for all terms (search terms and highlight terms) -------
+search_arguments$tresholds <- collect_treshold_values(raw_highlight_terms) %>%
+    .[highlight_terms_not_duplicated]
+search_arguments$custom_column <- collect_custom_column(raw_highlight_terms) %>%
+    .[highlight_terms_not_duplicated]
+
+# Checking search arguments -----------------------------------------------
 search_arguments$all_ok <- check_all_input()
