### 2019-04-25 #1342 @kbenoit
Sorry for the delay. I've now had time to examine the package and the paper in detail, and here are my comments.

### Overall

I really like the application and its simplicity. It looks great and is very functional. I had almost no problem installing it and testing it. I think this makes a nice addition to text analysis tools. However given the JOSS criteria I think it could be improved a in few ways that will likely not require too much work.

### General checks

The last GitHub release (0.4.0) is behind the current version (0.4.0.9000) but this can be corrected upon acceptance.

### Documentation

Yes although in its current form, it's very oriented toward documents that span dates, yet the statement of need in the paper speaks to digital humanities and other fields where this may not be the case. See my comments on this below.

Yes, but I had to install **PhantomJS** before I could get the tests to work. I suggest adding a note about this in the `README.md` on GitHub.

The Russia example is shown nicely in the GitHub README. But I think that another example that could excite digital humanities scholars would be to apply it to any corpus of documents chapters of a novel, such as _Moby Dick_ as it is analyzed in Jockers, M. L. (2014). _Text analysis with R for students of literature_. New York: Springer. (We replicate this for **quanteda** [here](http://quanteda.io/articles/pkgdown/replication/digital-humanities.html). I think that there are far more corpora that lack dates than that have them, so generalizing this and demonstrating it as an example would greatly broaden the user base of the package. Demonstrating the package on _Moby Dick_ would be a great application and it's easy to access that dataset online or bundle it with the package. (You would need to segment it by chapter first but this is not difficult.)

    I'd like to see a bit more in the documentation for the `prepare_data()` function, especially as relates to the structure of its variables such as Date and other document metadata fields.

### Software paper

    Yes but see my comments above relating to the requirement that documents have a date attached.

### Other

-  "paramater" and "treshold" are misspellings and should be corrected.
-  I think a better name would be **corpusexplorer**, since it only works with one corpus at a time (and this is easier to pronounce 😁)
-  I would really like to see `prepare_data()` defined as a generic with one method for an input `data.frame` to make it easy to extend this in the future. For example, this would then make it easy to add a method for a **quanteda** corpus in that package to make it easy to hand off a **quanteda** corpus to **corporaexplorer**.
-  Rather than require a date to be attached to each document, I think it would be better to replace this with an _optional_ sequence variable, and to assign one in the document order if none is given. If this is a date, great, and the package can use dates as is. Otherwise the sequence items would simply be serial numbers.
