### 2019-04-25 #1342 @kgjerde
@kbenoit Thank you very much indeed for the review and proposals! I will start pondering them this weekend.
