### 2019-04-26 0e1f819
prepare_data() defined as a generic with one method for an input data.frame
