+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(prepare_data,data.frame)
 export("%>%")
 export(prepare_data)
 export(run_corpus_explorer)
