### 2019-04-30 #1342 @leouieda
Thanks for the comments @kbenoit :+1:

@kgjerde please make sure to link back to the review issue in any issues or PRs opened in the software repo. Let us know when you have addressed the comments or if you want any clarification or further input.
