+++ README.Rmd
@@ -172,7 +172,7 @@ one needs to know is basically this:
 Thus, (in a case insensitive search):
 ``` r
 arctic  # will match both "Arctic" and "Antarctic"
-\barctic  # will match only "Arctic".
+\barctic  # will match only "Arctic"
 
 civili.ation  # will match both "civilisation" and "civilization"
 ```
