### 2019-05-01 163099d
Adding backwards compatibility re DATE_BASED_CORPUS constant
