+++ inst/download/global/backwards_compatibility.R
@@ -7,3 +7,7 @@ if ("mnd_vert" %in% colnames(abc$original_data$data_365)) {
         dplyr::rename(abc$original_data$data_dok,
                Text_original_case = Text_case)
 }
+
+if (is.null(abc$date_based_corpus)) {
+    abc$date_based_corpus <- TRUE
+}
