### 2019-05-01 5ed599b
Date_based vs not: sidebar ui - time filtering and plot mode
