+++ inst/explorer/ui/hide_ui_sidebar_plot_mode.R
@@ -1,0 +1,4 @@
+if (DATE_BASED_CORPUS == FALSE) {
+  shinyjs::hideElement(id = "modus")
+  shinyjs::hideElement(selector = ".conditional_hr")
+}
