### 2019-05-01 b5ab29a
Plot size
Improved, more flexible algorithm.
In function
prepared for both date and non-date corpora
