+++ inst/explorer/app.R
@@ -15,6 +15,7 @@ source("./global/function_filter_corpus.R", local = TRUE)
 source("./global/function_find_row_in_df.R", local = TRUE)
 source("./global/function_format_date.R", local = TRUE)
 source("./global/function_insert_doc_links.R", local = TRUE)
+source("./global/function_plot_size.R", local = TRUE)
 source("./global/function_visualise_corpus.R", local = TRUE)
 source("./global/function_visualise_document.R", local = TRUE)
 source("./global/functions_info.R", local = TRUE)
