### 2019-05-01 9e00ad2
info_plot to work with non-year x-axis (non-date-corpus)
