### 2019-05-01 101b9c8
Config accommodates non-date-corpora
