+++ inst/explorer/global/config_tests.R
@@ -6,11 +6,17 @@ library(corporaexplorer)
 
 loaded_data <- corporaexplorer::test_data
 
+source("./global/backwards_compatibility.R", local = TRUE)  # TODO should noe be ncessary in tests.
+
 # Constants ---------------------------------------------------------------
 
 source("./global/constants.R", local = TRUE)
 
 # From corporaexplorerobject --------------------------------------------
+
+DATE_BASED_CORPUS <- loaded_data$date_based_corpus
+# To decide if to display x-axis label in plot:
+ONLY_ONE_GROUP_IN_NON_DATE_BASED_CORPUS <- is.null(loaded_data$original_data$grouping_variable) & DATE_BASED_CORPUS == FALSE
 
 INFO_COLUMNS <- loaded_data$columns_for_info
 NO_MATRIX <- identical(loaded_data$ordvektorer$data_dok, FALSE)
