### 2019-05-01 9b5d72b
Wall view by default in non-date-corpora
