### 2019-05-01 17b2a50
New algorithm for wall view plot
