### 2019-05-01 f50ede3
Wall plot design
No longer thick lines between years
Instead distance + thin line between years/categories
