### 2019-05-01 4dbea25
Remove redundant variables
