+++ inst/explorer/global/corpus_plot_functions/plotting_corpus_day.R
@@ -37,9 +37,8 @@ plotting_corpus_day <-
                                xend = test1[test1$df == 1, ]$x_max,
                                y = test1[test1$df == 1, ]$y_min,
                                yend = test1[test1$df == 1, ]$y_max + linjer - 1,  # + 2 fordi tre rader i hver dokument FIX
-                               Year = test1$Year[test1$df == 1],
-                               Month = test1$Month[test1$df == 1])
-    
+                               Year = test1$Year[test1$df == 1])
+
     rect_tib_vertikal$x_mid = (rect_tib_vertikal$x + rect_tib_vertikal$xend)/2
     rect_tib_vertikal$y_mid = (rect_tib_vertikal$y + rect_tib_vertikal$yend)/2
     
