### 2019-05-01 8eb8a18
No axis label in plot if no real grouping variables
