### 2019-05-01 9b051f8
Update function_visualise_corpus.R
