+++ inst/explorer/global/function_visualise_corpus.R
@@ -43,17 +43,16 @@ visualiser_korpus <-
            modus) {
 
     linjer <- length(search_arguments$search_terms)
-    .data$Month <- lubridate::month(.data$Date)
-    
+
 # 1. Check if search contains search terms --------------------------------
-    
+
     if (identical(search_arguments$search_terms, "")) {
       .data$Term_1 <- NA
 
     } else{
 
 # 2. Count search term hits and assign factors for labelling --------------
-      
+
       .data <-
         count_search_terms_hits(.data,
                                 search_arguments,
@@ -70,14 +69,15 @@ visualiser_korpus <-
 
     if (modus == "data_365") {
      # width = 53 # fordi uker
+      .data$Month <- lubridate::month(.data$Date)
       .data$label_id <- seq_len(nrow(.data))
       .data <- create_coordinates_1_data_365(.data)
     } else if (modus == "data_dok") {
-      
+
       .data <- create_coordinates_1_data_dok(.data, linjer)
     } else if (modus == "day") {
 
-      .data <- create_coordinates_1_data_dok(.data, linjer, width = 15)
+      .data <- create_coordinates_1_data_dok(.data, linjer, max_width_for_row = 15)
     }
 
 # 4. Create plot coordinates, step 2 (if search terms > 1) ----------------
@@ -85,9 +85,9 @@ visualiser_korpus <-
     .data <-
       create_coordinates_several_search_terms(.data, linjer)
     .data <- dplyr::select(.data, -dplyr::starts_with("Term_"))
-    
+
 # 5. Create plot coordinates, step 3 (distance between days etc.) ---------
-   
+
     if (modus == "data_365") {
       .data <- create_distance_coordinates_365(.data,
                                                linjer,
@@ -105,20 +105,20 @@ visualiser_korpus <-
                                                 linjer,
                                                 day_distance)
     }
-    
+
 # 6. Label x and y axes ---------------------------------------------------
     y_text <- label_y_axis(.data)
-    
+
     if (modus == "data_365") {
       x_breaks <- label_x_axis_365(.data)
     }
-    
+
 # 7. Assign colours to plot labels/factors (up to 2 terms) ----------------
 
     temp_variable_for_unpacking <- colours_to_plot_and_legend(.data, linjer, number_of_factors)
     .data <- temp_variable_for_unpacking[[1]]  # colours_to_plot_and_legend(.data, linjer, number_of_factors)[[1]]
     til_legend <- temp_variable_for_unpacking[[2]] # colours_to_plot_and_legend(.data, linjer, number_of_factors)[[2]]
-    
+
 # 8. ggplotting -----------------------------------------------------------
 
     if (modus == "data_365") {
