### 2019-05-01 8f7bd30
Accommodate non-date-based-corpora
Mostly if statements
