+++ inst/explorer/server/2_event_search_button/UI_element_control.R
@@ -25,31 +25,33 @@ shinyWidgets::updateCheckboxGroupButtons(session,
 shinyjs::hideElement("edit_info_plot_legend_keys")
 
 # Update date inputs, making sure they remain within corpus date range
-if (search_arguments$time_filtering_mode == "Year range") {
-  updateDateRangeInput(
-    session,
-    "date_calendar",
-    start = if (as.Date(paste0(input$date_slider[1], "-01-01")) >
-      min(loaded_data$original_data$data_dok$Date)) {
-      as.Date(paste0(input$date_slider[1], "-01-01"))
-    } else {
-      min(loaded_data$original_data$data_dok$Date)
-    },
-    end = if (as.Date(paste0(input$date_slider[2], "-12-31")) <
-      max(loaded_data$original_data$data_dok$Date)) {
-      as.Date(paste0(input$date_slider[2], "-12-31"))
-    } else {
-      max(loaded_data$original_data$data_dok$Date)
-    }
-  )
-} else if (search_arguments$time_filtering_mode == "Date range") {
-  updateSliderInput(session,
-    "date_slider",
-    value = c(
-      lubridate::year(input$date_calendar[1]),
-      lubridate::year(input$date_calendar[2])
+if (DATE_BASED_CORPUS == TRUE) {
+  if (search_arguments$time_filtering_mode == "Year range") {
+    updateDateRangeInput(
+      session,
+      "date_calendar",
+      start = if (as.Date(paste0(input$date_slider[1], "-01-01")) >
+        min(loaded_data$original_data$data_dok$Date)) {
+        as.Date(paste0(input$date_slider[1], "-01-01"))
+      } else {
+        min(loaded_data$original_data$data_dok$Date)
+      },
+      end = if (as.Date(paste0(input$date_slider[2], "-12-31")) <
+        max(loaded_data$original_data$data_dok$Date)) {
+        as.Date(paste0(input$date_slider[2], "-12-31"))
+      } else {
+        max(loaded_data$original_data$data_dok$Date)
+      }
     )
-  )
+  } else if (search_arguments$time_filtering_mode == "Date range") {
+    updateSliderInput(session,
+      "date_slider",
+      value = c(
+        lubridate::year(input$date_calendar[1]),
+        lubridate::year(input$date_calendar[2])
+      )
+    )
+  }
 }
 
 # Fjerner faner (relevant for data_dok)
