### 2019-05-01 cdc9a7b
Fix doc box title when no grouping in non-date corpora
