### 2019-05-02 b560bac
Ensure correct "error" message in app
