### 2019-05-02 5dae9a4
Changed plot size algorithm
