+++ inst/explorer/global/constants.R
@@ -17,4 +17,4 @@ MAX_DOCS_IN_WALL_VIEW <- 12000
 # Constants for plot size and plot design ---------------------------------
 
 MAX_WIDTH_FOR_ROW <- 75  # Must be > 0
-X_FACTOR <- 20  # Multiplication element in function_plot_size()
+X_FACTOR <- 10  # Multiplication element in function_plot_size()
