### 2019-05-02 2c50ca6
Accommodate non-date corpora in info tab
