### 2019-05-04 45575d7
Added version field in corporaexplorerobject
