### 2019-05-04 5b204fa
Accommodate non-date corpora in extractor-app
