+++ inst/download/global/config_tests.R
@@ -6,6 +6,20 @@ library(corporaexplorer)
 
 abc <- test_data
 
+source("./global/backwards_compatibility.R", local = TRUE)
+
+DATE_BASED_CORPUS <- abc$date_based_corpus
+
+# Regarding titles for html report
+contains_grouping_variable <-
+  !is.null(abc$original_data$grouping_variable)
+if (contains_grouping_variable == TRUE) {
+GROUPING_VARIABLE <-
+    abc$original_data$grouping_variable
+} else {
+   GROUPING_VARIABLE <- NA
+}
+
 INFO_COLUMNS <- abc$columns_for_info
 abc <- abc$original_data$data_dok
 abc$Text <- abc$Text_original_case
