+++ inst/download/global/config.R
@@ -7,9 +7,20 @@ source("./global/backwards_compatibility.R", local = TRUE)
 
 if (class(abc) == "corporaexplorerobject") {
   INFO_COLUMNS <- abc$columns_for_info
+  DATE_BASED_CORPUS <- abc$date_based_corpus
+
+  # Regarding titles for html report
+  contains_grouping_variable <-
+      !is.null(abc$original_data$grouping_variable)
+  if (contains_grouping_variable == TRUE) {
+    GROUPING_VARIABLE <-
+        abc$original_data$grouping_variable
+  } else {
+       GROUPING_VARIABLE <- NA
+  }
+
   abc <- abc$original_data$data_dok
   abc$Text <- abc$Text_original_case
-
 }
 
 my_colours <-
