### 2019-05-05 16f1243
Differentiate between css for document box in data_dok and data_365
