+++ inst/explorer/server/4_event_day_map_click.R
@@ -1,8 +1,10 @@
 shiny::observeEvent(input$dag_klikk, {
     # Auto-scrolling to the top when new document clicked
-    shinyjs::runjs(sprintf("$('.boxed_doc').scrollTo('#%s');", 1))
+    shinyjs::runjs(sprintf("$('.boxed_doc_%s').scrollTo('#%s');",
+                           plot_mode$mode,
+                           1))
     shinyjs::runjs("$('.class_doc_box .nav-tabs-custom').scrollTo('0%');")
-    
+
     min_rad <-
         finn_min_rad(input$dag_klikk, session_variables$plotinfo_dag)
 
