+++ inst/explorer/server/5_event_document_visualisation_click.R
@@ -3,7 +3,8 @@ shiny::observeEvent(input$dok_vis_click, {
     musepos_dok_vis <- input$dok_vis_click
     rute_nr <- ceiling(musepos_dok_vis$x - 0.5)
     shinyjs::runjs(sprintf(
-        "$('.boxed_doc').scrollTo('#%s', duration = 200);",
+        "$('.boxed_doc_%s').scrollTo('#%s', duration = 200);",
+        plot_mode$mode,
         rute_nr
     ))
 })
