### 2019-05-05 52df8aa
plotsize step size in sidebar
