### 2019-05-05 75c4831
Add prepare_data() method for character vector
