+++ man/prepare_data.character.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/prepare_data.R
+\name{prepare_data.character}
+\alias{prepare_data.character}
+\title{Quickly explore character vector}
+\usage{
+\method{prepare_data}{character}(dataset, ...)
+}
+\arguments{
+\item{dataset}{A non-empty character vector.}
+
+\item{...}{Other arguments to be passed to \code{prepare_data}.}
+}
+\value{
+A corporaexplorerobject.
+}
+\description{
+Quick convertion of character vector to simple corporaexplorerobject
+  with no metadata.
+}
+\examples{
+alphabet_corpus <- prepare_data(LETTERS)
+
+\dontrun{
+# Running exploration app:
+run_corpus_explorer(alphabet_corpus)
+}
+}
