### 2019-05-05 b9f2d8d
Changes in prepare_data()
- Now has possibility to create non-date-based corpora
- Included arguments for non-date-based corpora: data_based_corpus (logical) and grouping_variable.
- A bit more documentation
- Removed argument columns_to_include, as really not very useful
- Some new argument checking
