+++ R/prepare_data.R
@@ -5,64 +5,53 @@
 #' @param df Data frame with Date column (Date), Text column (character), and
 #'   optionally Title (character), URL (character), and Type (character)
 #'   columns.
-#' @param columns_to_include Character vector. The columns from df to be
-#'   included in corpus object, in addition to the required Date and Text. E.g.
-#'   \code{c("Title", "URL")}. If NULL, the default, all columns will be
-#'   included.
 #' @param normalise Should non-breaking spaces (U+00A0) and soft hyphens
 #'   (U+00ad) be normalised?
 #' @return A tibble ("data_dok")
 #' @keywords internal
-transform_regular <- function(df, columns_to_include = NULL, normalise = TRUE) {
+transform_regular <- function(df, normalise = TRUE) {
   message("Starting.")
 
-  if (is.null(columns_to_include)) {
-    new_df <- df
-  } else {
-    new_df <-
-      dplyr::select(df, Date, Text, columns_to_include) # Velg kolonner her.
-  }
-
   if (normalise == TRUE) {
-    for (i in seq_along(new_df)) {
+    for (i in seq_along(df)) {
       # Only for character columns
-      if (is.character(new_df[[i]])) {
+      if (is.character(df[[i]])) {
         ## På grunn av non-breaking-spaces:
-        new_df[[i]] <- gsub("\u00A0", " ", new_df[[i]], fixed = TRUE)
+        df[[i]] <- gsub("\u00A0", " ", df[[i]], fixed = TRUE)
         ## And "soft hyphens"
-        new_df[[i]] <- gsub("\u00ad", "", new_df[[i]], fixed = TRUE)
+        df[[i]] <- gsub("\u00ad", "", df[[i]], fixed = TRUE)
       }
     }
   }
 
-  new_df$Year <- lubridate::year(new_df$Date)
+  df$Year <- lubridate::year(df$Date)
 
-  new_df$Tile_length <- nchar(new_df$Text) %>%
+  df$Tile_length <- nchar(df$Text) %>%
     scales::rescale(to = c(1, 10), from = c(0, max(.)))
 
-  new_df$ID <- seq_len(nrow(new_df))
+  df$ID <- seq_len(nrow(df))
 
-  new_df$Text_original_case <- new_df$Text
+  df$Text_original_case <- df$Text
 
-  new_df$Text <-
-    stringr::str_to_lower(new_df$Text) # for søke-purposes
+  df$Text <-
+    stringr::str_to_lower(df$Text) # for søke-purposes
 
-  new_df <- dplyr::select(new_df,
+  df <- dplyr::select(df,
                 ID,
                 Date,
                 Text,
                 Text_original_case,
                 Tile_length,
                 Year,
-                sort(colnames(new_df)[!colnames(new_df) %in% c("ID",
+                sort(colnames(df)[!colnames(df) %in% c("ID",
                                                                "Date",
                                                                "Text",
                                                                "Text_original_case",
                                                                "Tile_length",
                                                                "Year")]))
 
-  message("1/3 Document data frame done.")
-  return(new_df)
+  message("Document data frame done.")
+  return(df)
 }
 
 
@@ -161,7 +150,7 @@ transform_365 <- function(new_df) {
                           Tile_length
                           )
 
-  message("2/3 Calendar data frame done.")
+  message("Calendar data frame done.")
   return(df_365)
 }
 
@@ -216,7 +205,7 @@ matrix_via_r <- function(df, matrix_without_punctuation = TRUE) {
     stringr::str_replace_all(" {2,}", " ") %>%
     stringr::str_trim()
 
[…]
