### 2019-05-06 7e98dda
Adding 'grouping_variable' argument check
