### 2019-05-06 707ddde
Adding Bible and Jane Austen examples
- and including examples folder in .Rbuildignore
