### 2019-05-06 42db195
namespace add method
