### 2019-05-06 b3fdd69
Add info about PhantomJS to DESCRIPTION
