### 2019-05-06 ff6e66a
Documentation of prepare_data()
