### 2019-05-06 b6dcd95
Updated and restructured README
- Mainly to include non-date-based corpora
