+++ README.md
@@ -18,9 +18,36 @@ corpora/text collections through a `Shiny` app graphical user interface.
 
 **corporaexplorer**’s intended primary audience is qualitatively
 oriented researchers who rely on close reading of textual documents as
-part of their academic activity, and for whom a typical workflow often
-includes analysing a set of texts downloaded from a website based on
-keyword search in the website’s internal search engine.
+part of their academic activity, but should also be a useful supplement
+for those doing quantitative textual research and wishing to visit the
+texts under study. Finally, by offering a convenient way to explore any
+character vector, it can also be useful for a wide range of other R
+users.
+
+While collecting and preparing the text collections to be explored
+requires some R programming knowledge, using the Shiny app(s) for
+exploring and extracting documents from the corpus should be rather
+intuitive also for someone with no programming knowledge – when first
+set up by a collaborator. Thus, the ambition is that the package should
+be useful for anyone with a rudimentary knowledge of R – or who knows
+anyone with this knowledge.
+
+## Installation
+
+To install **corporaexplorer**, run the following from an R console:
+
+``` r
+install.packages("devtools")
+devtools::install_github("kgjerde/corporaexplorer")
+```
+
+## Example use cases and example corpora
+
+### A. Date-based corpora
+
+For many qualitatively oriented researchers, a typical workflow includes
+analysing a set of texts downloaded from a website based on keyword
+search in the website’s internal search engine.
 
 **corporaexplorer** aims to enhance such workflows in several ways. The
 package makes it easier to interact with larger text collections than
@@ -44,57 +71,6 @@ Example corpora and use cases include:
     of newspaper articles etc. retrieved from a database.
   - **Personal text collections**: any collection of texts that are
     meaningful to explore with a time aspect.
-  - **Customised corpora**, e.g. text collections with sentences instead
-    of documents as base unit.
-
-While collecting and preparing the text collections to be explored
-requires some R programming knowledge, using the Shiny app(s) for
-exploring and extracting documents from the corpus should be rather
-intuitive also for someone with no programming knowledge – when first
-set up by a collaborator.
-
-Thus, the ambition is that the package should be useful for anyone with
-a rudimentary knowledge of R – or who knows anyone with this knowledge.
-
-## Installation
-
-To install **corporaexplorer**, run the following from an R console:
-
-``` r
-install.packages("devtools")
-devtools::install_github("kgjerde/corporaexplorer")
-```
-
-## A note on platforms and encoding
-
-**corporaexplorer** works on Mac OS, Windows and
-Linux,<sup>[1](#footnote1)</sup> and there are some important
-differences in how R handles text on the different platforms. If you are
-working with plain English text, there will most likely be no issues
-with encoding on any platform. Unfortunately, working with
-non-[ASCII](https://en.wikipedia.org/wiki/ASCII) encoded text in R
-(e.g. non-English characters), *can* be complicated – in particular on
-Windows.
-
-**On Mac OS or Linux**, problems with encoding will likely not arise at
-all. If problems do arise, they can typically be solved by making the R
-“locale” unicode-friendly (e.g. `Sys.setlocale("LC_ALL",
-"en_US.UTF-8")`). NB\! This assumes that the text is UTF-8 encoded, so
-if changing the locale in this way does not help, make sure that the
-text is encoded as UTF-8 characters. Alternatively, if you can ascertain
-the character encoding, set the locale correspondingly.
-
-**On Windows**, things can be much more complicated. The most important
-thing is to check carefully that the texts appear as expected in
-`corporaexplorer`’s apps, and that the searches function as expected. If
[…]
