### 2019-05-06 kgjerde/corporaexplorer#12 @kgjerde
Review: prepare_data() as generic
> I would really like to see prepare_data() defined as a generic with one method for an input data.frame to make it easy to extend this in the future. For example, this would then make it easy to add a method for a quanteda corpus in that package to make it easy to hand off a quanteda corpus to corporaexplorer.

Implemented in kgjerde/corporaexplorer@0e1f819

This is the first time I have defined generics, so please let me know if this is not what you had in mind.
