### 2019-05-06 kgjerde/corporaexplorer#13 @kgjerde
Review: misspellings
> "paramater" and "treshold" are misspellings and should be corrected.

Fixed in kgjerde/corporaexplorer@7ac8561
