### 2019-05-06 kgjerde/corporaexplorer#14 @kgjerde
Review: PhantomJS
> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
>> Yes, but I had to install PhantomJS before I could get the tests to work. I suggest adding a note about this in the README.md on GitHub.

This appears to be a shinytest issue, and they claim that they provide a useful error message as well as function to install PhantomJS (https://github.com/rstudio/shinytest/issues/175).

I am a bit afraid of further cluttering the README, so I have added a note in DESCRIPTION in commit kgjerde/corporaexplorer@b3fdd69. Please let me know if you think this is not sufficient.
