### 2019-05-06 d0bb331
README footnotes
