+++ README.Rmd
@@ -73,12 +73,12 @@ A more typical corpus can be downloaded for testing
 [here](https://ndownloader.figshare.com/files/14626970?private_link=820c105c12755d853801)
 -- this is a "corporaexplorerobject" based on a document collection scraped from the Russian governmental website kremlin.ru (7751 documents, more info [here](https://figshare.com/s/820c105c12755d853801)).
 The equivalent material in Russian (10,221 documents) can be downloaded [here](https://figshare.com/s/be0d2cfa195e43160fcb).
-A third, somewhat bigger (but less carefully prepared) test corpus (26,654 documents) from the same website is available [here](https://figshare.com/s/fcd93f6f93bb23be3bb1).<sup>[2](#footnote2)</sup>
+A third, somewhat bigger (but less carefully prepared) test corpus (26,654 documents) from the same website is available [here](https://figshare.com/s/fcd93f6f93bb23be3bb1).<sup>[1](#footnote1)</sup>
 
 
 ### B. Text collections that are not organised by dates
 
-There are far more corpora that lack dates than that have them,
+There are far more corpora that lack dates than that have them,<sup>[2](#footnote2)</sup>
 and it is simple to create a "corporaexplorerobject" organised around
 for example chapters in books or a different structure -- or just as a sequence of texts.
 
@@ -317,7 +317,7 @@ are carried out as full text searches with `stringr`. Again, note that a single 
 
 ## A note on platforms and encoding
 
-**corporaexplorer** works on Mac OS, Windows and Linux,<sup>[1](#footnote1)</sup>
+**corporaexplorer** works on Mac OS, Windows and Linux,<sup>[4](#footnote4)</sup>
 and there are some important differences in how R handles text
 on the different platforms.
 If you are working with plain English text,
@@ -357,6 +357,7 @@ Contributions in the form of feedback, bug reports and code are most welcome. Wa
 
 <hr>
 
-<a name="footnote1">*1</a>. That said, the package is developed on Mac OS, and the Shiny apps' user interface is a bit clunkier on Windows.*  
-<a name="footnote2">*2</a>. All this material is licensed under Creative Commons Attribution 4.0 International (CC BY 4.0).*  
-<a name="footnote3">*3</a>. Using the [`jquery.scrollTo`](https://github.com/flesler/jquery.scrollTo) plugin.* 
+<a name="footnote1">*1</a>. All this material is licensed under Creative Commons Attribution 4.0 International (CC BY 4.0).*  
+<a name="footnote2">*2</a>. Thanks to Kenneth Benoit for pointing this out.*  
+<a name="footnote3">*3</a>. Using the [`jquery.scrollTo`](https://github.com/flesler/jquery.scrollTo) plugin.*  
+<a name="footnote4">*4</a>. That said, the package is developed on Mac OS, and the Shiny apps' user interface is rather clunkier on Windows.*
