### 2019-05-06 kgjerde/corporaexplorer#16 @kgjerde
Review: Name
> I think a better name would be corpusexplorer, since it only works with one corpus at a time (and this is easier to pronounce 😁)

Fair point, but I think I still would like to stick with corporaexplorer, partly because it is more "googleable".
