### 2019-05-06 kgjerde/corporaexplorer#17 @kgjerde
Review: version
>  Version: Does the release version given match the GitHub release (v0.4.0)?
>> The last GitHub release (0.4.0) is behind the current version (0.4.0.9000) but this can be corrected upon acceptance.

Yes, I will make sure to change version number before archiving the repo to Zenodo or similar.
