### 2019-05-06 kgjerde/corporaexplorer#18 @kgjerde
Review: documentation
> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>> I'd like to see a bit more in the documentation for the prepare_data() function, especially as relates to the structure of its variables such as Date and other document metadata fields.

Please have a look at the updated prepare_data documentation and README.
