### 2019-05-06 d655730
README wording [skip ci]
