+++ README.Rmd
@@ -79,8 +79,10 @@ A third, somewhat bigger (but less carefully prepared) test corpus (26,654 docum
 ### B. Text collections that are not organised by dates
 
 There are far more corpora that lack dates than that have them,<sup>[2](#footnote2)</sup>
-and it is simple to create a "corporaexplorerobject" organised around
-for example chapters in books or a different structure -- or just as a sequence of texts.
+and it is simple to create a "corporaexplorerobject"
+from a corpus with a different structure,
+for example a collection of books, each organised in chapters
+-- or just from a sequence of texts.
 
 See [here](examples/bible.md) for an example of how **corporaexplorer** can be used to explore the Bible,
 and [here](examples/jane_austen.md) for a starter on how to explore Jane Austen's novels.
