### 2019-05-06 #1342 @kgjerde
@kbenoit

I have now addressed or commented your points summarised in:
* kgjerde/corporaexplorer#12  (prepare_data() as generic)
* kgjerde/corporaexplorer#13  (misspellings)
* kgjerde/corporaexplorer#14  (PhantomJS)
* kgjerde/corporaexplorer#15  (several points regarding corpora without dates)
* kgjerde/corporaexplorer#16  (name)
* kgjerde/corporaexplorer#17  (versioning)
*  kgjerde/corporaexplorer#18  (documentation)

Thank you again for very useful input. In particular, allowing for corpora without dates demanded some work on the app internals, but you are clearly correct that this significantly broadens the potensial use cases. I look forward to hearing what you think about the changes made.
