+++ README.Rmd
@@ -103,7 +103,7 @@ documents from a "corporaexplorerobject" in a reading-friendly format.
 
 The `prepare_data()` function returns a "corporaexplorerobject" that can be explored in the package's two Shiny apps.
 
-The two most important arguments are:
+The three most important arguments are:
 
 * `dataset`: a data frame with, as a minimum, a `Text` column. If `date_based_corpus` is `TRUE` (the default), dataset must also contain a column "Date" (of class Date).
 * `date_based_corpus`. Default is `TRUE`. Set to `FALSE` if the corpus is not to be organised according to document dates.
