### 2019-05-06 0a54c4e
readme wording [skip ci]
