### 2019-05-06 0ffc905
Wording in Bible example [skip ci]
