### 2019-05-06 633ed7d
wording in Bible example [skip ci]
