### 2019-05-06 46f503c
Made navigation link class less prone for accidental highlighting
