+++ README.Rmd
@@ -80,7 +80,7 @@ A third, somewhat bigger (but less carefully prepared) test corpus (26,654 docum
 ### B. Text collections that are not organised by date
 
 There are far more corpora that lack dates than that have them,<sup>[2](#footnote2)</sup>
-and it is simple to create a "corporaexplorerobject"
+and it is easy to create a "corporaexplorerobject"
 from a corpus with a different structure,
 for example a collection of books, each organised in chapters
 -- or just from a sequence of texts.
