+++ inst/explorer/ui/css_js_import.R
@@ -1,5 +1,5 @@
 shiny::tags$head(
-    shiny::includeCSS("js_css/app_css.css"),
+    shiny::includeCSS("js_css/explorer_css.css"),
     shiny::includeScript("js_css/box_width.js"),
     shiny::includeScript("js_css/enter_search.js"),
     shiny::includeScript("js_css/jquery.scrollTo.js")
