### 2019-05-07 4bedd53
Hover info also for non-date-based corpora
- Abstracted doc title/info rendering into function because DRY
