### 2019-05-07 bcd80dc
Avoid wrong tooltip position in some cases with very small plots
