### 2019-05-07 a30369c
Removed class from navigation link (ref. 46f503c)
