+++ README.Rmd
@@ -79,7 +79,7 @@ A third, somewhat bigger (but less carefully prepared) test corpus (26,654 docum
 
 ### B. Text collections that are not organised by date
 
-There are far more corpora that lack dates than that have them,<sup>[2](#footnote2)</sup>
+Not all corpora are organised by date,
 and it is easy to create a "corporaexplorerobject"
 from a corpus with a different structure,
 for example a collection of books, each organised in chapters
@@ -249,7 +249,7 @@ knitr::include_graphics("man/figures/day_corpus.png")
 
 <br>
 
-Clicking on a "document tile" produces two things. First, the full text of the document with search terms highlighted. Second, above the text a tile chart consisting of n tiles where each tile represents a 1/n part of the document, and where the colour in a tile indicates whether and how many times the search term is found in that part of the document. Clicking on a tile scrolls the document to the corresponding part of the document.<sup>[3](#footnote3)</sup>
+Clicking on a "document tile" produces two things. First, the full text of the document with search terms highlighted. Second, above the text a tile chart consisting of n tiles where each tile represents a 1/n part of the document, and where the colour in a tile indicates whether and how many times the search term is found in that part of the document. Clicking on a tile scrolls the document to the corresponding part of the document.<sup>[2](#footnote2)</sup>
 
 ```{r out.width="80%", echo=FALSE}
 knitr::include_graphics("man/figures/wall.png")
@@ -320,7 +320,7 @@ are carried out as full text searches with `stringr`. Again, note that a single 
 
 ## A note on platforms and encoding
 
-**corporaexplorer** works on Mac OS, Windows and Linux,<sup>[4](#footnote4)</sup>
+**corporaexplorer** works on Mac OS, Windows and Linux,<sup>[3](#footnote3)</sup>
 and there are some important differences in how R handles text
 on the different platforms.
 If you are working with plain English text,
@@ -361,6 +361,5 @@ Contributions in the form of feedback, bug reports and code are most welcome. Wa
 <hr>
 
 <a name="footnote1">*1</a>. All this material is licensed under Creative Commons Attribution 4.0 International (CC BY 4.0).*  
-<a name="footnote2">*2</a>. Thanks to Kenneth Benoit for pointing this out.*  
-<a name="footnote3">*3</a>. Using the [`jquery.scrollTo`](https://github.com/flesler/jquery.scrollTo) plugin.*  
-<a name="footnote4">*4</a>. That said, the package is developed on Mac OS, and the Shiny apps' user interface is rather clunkier on Windows.*
+<a name="footnote2">*2</a>. Using the [`jquery.scrollTo`](https://github.com/flesler/jquery.scrollTo) plugin.*  
+<a name="footnote3">*3</a>. That said, the package is developed on Mac OS, and the Shiny apps' user interface is rather clunkier on Windows.*
