### 2019-05-08 7cfda3f
Revert "Removed class from navigation link (ref. 46f503c)"
This reverts commit a30369c417f1ac60c7ce23c343b51f14d6194878.
