### 2019-05-08 cb9ac14
Corrected rects around dates/docs with more than one search term
- + cleaned up a bit
