+++ inst/explorer/global/corpus_plot_functions/plotting_corpus_data_365.R
@@ -1,15 +1,15 @@
 #' Main function for ggplot2 corpus map creation: calendar view
 #'
 #' @param test1 Data frame with columns XXXX
-#' @param x_breaks 
-#' @param y_text 
-#' @param til_legend 
-#' @param linjer 
+#' @param x_breaks
+#' @param y_text
+#' @param til_legend
+#' @param linjer
 #'
 #' @return A ggplot2 plot object
 plotting_corpus_data_365 <- function(test1, x_breaks, y_text, til_legend, linjer) {
     a <- ggplot2::ggplot()
-    
+
     a <- a +
         ggplot2::scale_y_reverse(
             breaks = y_text$breaks,
@@ -33,7 +33,7 @@ plotting_corpus_data_365 <- function(test1, x_breaks, y_text, til_legend, linjer
             axis.text = ggplot2::element_text(size=10)
         ) +
         ggplot2::guides(fill = ggplot2::guide_legend(ncol = 8, byrow = TRUE))
-    
+
     rect_tib_vertikal <- tibble::tibble(
         x = test1[test1$df == 1,]$x_min,
         xend = test1[test1$df == 1,]$x_max,
@@ -42,7 +42,7 @@ plotting_corpus_data_365 <- function(test1, x_breaks, y_text, til_legend, linjer
         Year = test1$Year[test1$df == 1],
         Month = test1$Month[test1$df == 1]
     )
-    
+
     rect_tib_vertikal$x_mid = (rect_tib_vertikal$x + rect_tib_vertikal$xend) /
         2
     rect_tib_vertikal$y_mid = (rect_tib_vertikal$y + rect_tib_vertikal$yend) /
@@ -60,7 +60,7 @@ plotting_corpus_data_365 <- function(test1, x_breaks, y_text, til_legend, linjer
             na.value = "white"
         )
     # }
-    
+
     # Dette er fyll basert på søk
     a <-
         a + ggplot2::geom_rect(data = test1[test1$df == 1,],
@@ -71,61 +71,7 @@ plotting_corpus_data_365 <- function(test1, x_breaks, y_text, til_legend, linjer
                           ymax = y_max,
                           fill = Term_color
                       ))
-    
-    
-    #    Dette er selve rutenettet
-    a <- a + ggplot2::geom_rect(
-        data = rect_tib_vertikal
-        ,
-        ggplot2::aes(
-            xmin = x,
-            ymin = y,
-            xmax = xend,
-            ymax = yend
-        ),
-        fill = NA,
-        color = "black",
-        size = 0.15
-    )# +
-    #  ggplot2::theme(legend.position = "none")
-    
-    # # Og her er for å indikere dager uten dokumenter:
-    a <-
-        a + ggplot2::geom_rect(
-            data = rect_tib_vertikal[rect_tib_vertikal$Day_without_docs == TRUE,],
-            ggplot2::aes(
-                xmin = x,
-                ymin = y,
-                xmax = xend,
-                ymax = yend
-            ),
-            fill = "white",
-            colour = "white",
-            size = 0.2,
-            na.rm = TRUE
-        )
-    
-    # Og her skjuler vi ikke-eksisterende dager og dager utenfor datasettet
-    a <-
-        a + ggplot2::geom_rect(
-            data = test1[test1$Invisible_fake_date == TRUE,],
-            ggplot2::aes(
-                xmin = x_min,
-                xmax = x_max,
-                ymin = y_min,
-                ymax = y_max
-            ),
-            fill = "white",
-            size = 1,
-            colour = "white"
-        )
-    
-    a <- a + ggplot2::scale_alpha_identity()
-    
-    
-    
-    
-    
+
[…]
