### 2019-05-08 d0b2214
Tweaked plot size algorithm to improve sizing when multiple search terms
