### 2019-05-09 725c5a0
Fix plot at startup in line with changes in cb9ac148
