### 2019-05-09 #1342 @trinker
The items brought up were addressed and have been checked off

> @kgjerde thank you for reporting back to us 👍
>
> 👋 @trinker @kbenoit please take another look at the project and paper when you have time. Here are a few things to keep in mind:
>
> * Have all your comments/concerns been addressed?
> * Are there still any items from the checklist at the top of this issue left un-checked?
>
> If you're satisfied with the revisions, please let me know and check any remaining items from the checklist.
