### 2019-05-10 #1342 @kbenoit
This all looks great to me, and 👏 for the clear and elegant use of issues in your repo with linked commit references. The generalization away from just dates is great, and the two examples are perfect here. Nice!  And the name is your decision of course.

One thing only I would suggest, a 20-second change:
* I don't think it would clutter the README at all to add a note about PhantomJS to the section on Installation. (It's more appropriate there than in DESCRIPTION, but you can leave it in DESCRIPTION too.)
