### 2019-05-10 c370f18
Add info about PhantomJS in README (#14)
