### 2019-05-10 kgjerde/corporaexplorer#14 @kgjerde
@kbenoit, is something like c370f18 what you had in mind?
