### 2019-05-10 #1342 @kgjerde
Excellent, thank you! @kbenoit, I have no addressed your last point (see kgjerde/corporaexplorer#14).

@leouieda I will now also take a new look at the paper itself, bearing in mind the changes I have made to the package in response to the reviews. Will post here when done.
