### 2019-05-10 #1342 @leouieda
:confetti_ball: @kbenoit @trinker thank you for your input! @kbenoit I see some items left unchecked in your section at the very top. Is this intentional?

@kgjerde OK, let us know when you're done and we'll compile a new version of the pdf so we can have a final round of reviews of the paper.
