### 2019-05-12 #1342 @kbenoit
Sorry, I should have checked those. I have now done so for everything except the DOI, since I have not seen the update paper yet.
