### 2019-05-14 e091e00
Change minimum size of plot
