### 2019-05-14 34a4d01
Improved insert_doc_navigation_links()
- Now inserted only where spaces
