+++ examples/bible.Rmd
@@ -36,7 +36,7 @@ bible <- paste(bible, collapse = "\n")
 # Identifying the beginning and end of the Bible / stripping PJ metadata
  # (technique borrowed from https://quanteda.io/articles/pkgdown/replication/digital-humanities.html).
 start_v <- stri_locate_first_fixed(bible, "The First Book of Moses:  Called Genesis")[1]
-end_v <- stri_locate_last_fixed(bible, "Amen.")[1]
+end_v <- stri_locate_last_fixed(bible, "Amen.")[2]
 bible <- stri_sub(bible, start_v, end_v)
 
 # In the file, every book in the bible is preceded by five newlines,
