### 2019-05-15 263e6d5
Thicker lines between years/groups if more search terms
