+++ inst/explorer/global/corpus_plot_functions/plotting_corpus_data_dok.R
@@ -104,7 +104,7 @@ if (length(coordinates_for_dividing_lines_btwn_groups) > 0) {
       "segment",
       linetype = "solid",
       colour = "lightgray",
-      size = 0.3,
+      size = 0.3 * linjer,
       x = min(test1$x_min) - (difference_min_max_x / 35),
       xend = max(test1$x_max) + (difference_min_max_x / 125),
       y = coordinates_for_dividing_lines_btwn_groups,
