### 2019-05-15 d5b3d99
Improved doc wall when several search terms
Vertical distance -> easier to distinguish between documents
