### 2019-05-17 b0a495c
Update paper
- slight revisions
- language edits
- added figure
