+++ paper.md
@@ -1,10 +1,11 @@
 ---
 title: 'corporaexplorer: an R package for dynamic exploration of text collections'
 tags:
-  - R
   - qualitative research
+  - mixed methods
   - text analysis
   - corpora
+  - R
   - Shiny
 authors:
   - name: Kristian Lundby Gjerde
@@ -13,30 +14,36 @@ authors:
 affiliations:
  - name: Research Fellow, Norwegian Institute of International Affairs (NUPI)
    index: 1
-date: 17 March 2019
+date: 17 May 2019
 bibliography: joss_bib.bib
 ---
 
 # Background
 
-Computer technology has profoundly changed the possibilities for doing research with text as data. Terms such as 'computational social science' [see @comp] and 'digital humanities'  [see @digitalhumanities] are now firmly established, indicative of how new methods have transformed whole fields of scholarly enquiry, enabling researchers to do things virtually unimaginable even a few years ago. For good reasons – computers excel in quantitative tasks – these advances have often been predominantly quantitatively oriented, and geared towards 'big data' of various size. Yet, developments in software and hardware have also opened up vast fields of immensely helpful -- and often unused -- possibilities for a great number of scholars engaged in various forms of *qualitative* text analysis.
+Computer technology has profoundly changed the possibilities for doing research with text as data. Terms such as ‘computational social science’ (see @comp) and ‘digital humanities’ (see @digitalhumanities) are now firmly established, indicating how new methods have transformed entire fields of scholarly enquiry, enabling researchers to do things unimaginable even a few years ago. For good reasons -- computers excel in quantitative tasks -- these advances have been predominantly quantitatively oriented, geared towards ‘big data’ of various sizes. However, developments in software and hardware have also opened up vast fields of immensely helpful -- and often untapped -- possibilities for many scholars engaged in various forms of *qualitative* text analysis.
 
 # The ``corporaexplorer`` R package  
 
-``corporaexplorer`` is an R package that uses the ``Shiny`` GUI (graphical user interface) framework for dynamic exploration of text collections, e.g. a collection of documents scraped from a governmental website or newspaper articles retrieved from a database. The intended primary audience is qualitatively oriented researchers in the social sciences and humanities who rely on close reading of textual documents as part of their academic activity -- and for whom a typical workflow often includes analysing a set of texts downloaded from a website based on keyword search in the website's internal search engine.
+``corporaexplorer`` is an R package that uses the ``Shiny`` GUI (graphical user interface) framework for dynamic exploration of text collections. The package is designed for use with a wide range of text collections; one example could be a collection of tens of thousands of documents scraped from a governmental website; another example could be the collected works of a novelist; a third example could be the chapters of a single book.
 
-``corporaexplorer``'s main goal is to facilitate more powerful, transparent, rigid, and efficient workflows in these kinds of qualitative research. It aims to help researchers to conduct open-ended explorations of text collections, and to identify and extract texts for further scrutiny -- and to do this in a fast and user-friendly manner. The hope is also to encourage mixed methods in text analysis, by using ``corporaexplorer`` in  combination with the excellent R packages that are available for quantitative text analysis [for a highly useful review, see @welbers], and to encourage (where licenses so permit) sharing of text collections.
[…]
