+++ joss_bib.bib
@@ -4,7 +4,7 @@
     year = {2019},
     url = {http://www.gagolewski.com/software/stringi/},
   }
-  
+
   @Manual{stringr,
     title = {{{s}}tringr: Simple, Consistent Wrappers for Common String Operations},
     author = {Hadley Wickham},
@@ -12,7 +12,7 @@
     note = {R package version 1.4.0},
     url = {https://CRAN.R-project.org/package=stringr},
   }
-  
+
     @Manual{shiny,
     title = {{{s}}hiny: Web Application Framework for {{R}}},
     author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
@@ -20,16 +20,17 @@
     note = {R package version 1.2.0},
     url = {https://CRAN.R-project.org/package=shiny},
   }
-  
+
     @Book{ggplot2,
     author = {Hadley Wickham},
     title = {{{g}}gplot2: Elegant Graphics for Data Analysis},
-    publisher = {Springer-Verlag New York},
+    publisher = {Springer-Verlag},
+    address = {New York},
     year = {2016},
     isbn = {978-3-319-24277-4},
     url = {http://ggplot2.org},
   }
-  
+
     @Manual{tidyverse,
     title = {{{t}}idyverse: Easily Install and Load the 'Tidyverse'},
     author = {Hadley Wickham},
@@ -37,7 +38,7 @@
     note = {R package version 1.2.1},
     url = {https://CRAN.R-project.org/package=tidyverse},
   }
-  
+
     @Manual{data.table,
     title = {{{d}}ata.table: Extension of `data.frame`},
     author = {Matt Dowle and Arun Srinivasan},
@@ -45,7 +46,7 @@
     note = {R package version 1.12.0},
     url = {https://CRAN.R-project.org/package=data.table},
   }
-  
+
     @Manual{re2r,
     title = {{{r}}e2r: RE2 Regular Expression},
     author = {{Qin Wenfeng}},
@@ -53,16 +54,16 @@
     note = {R package version 0.2.0},
     url = {https://CRAN.R-project.org/package=re2r},
   }
-  
+
     @Manual{R,
     title = {R: A Language and Environment for Statistical Computing},
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
-    address = {Vienna, Austria},
+    address = {Vienna},
     year = {2018},
     url = {https://www.R-project.org/},
   }
-  
+
   	@article{welbers,
 	langid = {english},
   	title = {Text {{Analysis}} in {{R}}},
@@ -97,7 +98,7 @@
 }
 
 	@book{digitalhumanities,
-	address = {Minneapolis, MN London},
+	address = {Minneapolis, MN},
 	title = {Debates in the digital humanities: 2016},
 	isbn = {978-0-8166-9954-4 978-0-8166-9953-7},
 	shorttitle = {Debates in the digital humanities},
