### 2019-05-17 #1342 @kgjerde
@leouieda I have now updated the paper (mainly in kgjerde/corporaexplorer@b0a495c, then just a few corrections). The paper is slightly revised (most of the diffs are language edits) and I have also added a figure to illustrate the apps.
