+++ paper.md
@@ -1,8 +1,5 @@
 ---
 title: 'corporaexplorer: an R package for dynamic exploration of text collections'
-output:
-  pdf_document:
-    latex_engine: xelatex
 tags:
 - qualitative research
 - mixed methods
