### 2019-05-17 #1342 @kgjerde
@leouieda Sorry about the repeated article rendering. There is just a small problem with unicode characters in the caption to figure 1, which is probably linked with whedon's pandoc pdf rendering, but I do not know how to fix it.
