### 2019-05-18 #1342 @leouieda
Hi @kgjerde, no worries. Compiling the PDF can be troublesome. I don't know why it's not rendering the unicode in the caption when it has no problem with other characters in the text. I'll have to check with some higher-ups.

@trinker and @kbenoit in the mean time, could you please go over the paper one final time? Please let me know if you're ready to sign off on this submission. Thank you for all your input!
