### 2019-05-18 9bca603
Remove Cyrillic from caption [skip ci]
