### 2019-05-19 ff25cb9
Add horizontal distance btwn tiles when multiple search terms
- For readability
- In doc wall view
+ plot size tweaking
