### 2019-05-20 bcd6e57
Small language edit in paper.
