### 2019-05-20 502f274
Clicking outside of tiles in corpus map no longer leads to autoscroll doc to top
- fixes small bug introduced in 16f1243
