+++ inst/explorer/server/3_event_corpus_map_click/js_auto_scroll.R
@@ -1,0 +1,5 @@
+# Auto scroll to top when new tile clicked --------------------------------
+shinyjs::runjs(sprintf("$('.boxed_doc_%s').scrollTo('#%s');",
+                       plot_mode$mode,
+                       1))
+shinyjs::runjs(sprintf("$('#dag_kart').scrollTo('%s');", 0))
