### 2019-05-20 580b215
Delete redundant file
- this function resides in create_factors_for_labelling.R
