### 2019-05-21 75765f0
Paper figure placement attempt
- LaTeX directly in md
