### 2019-05-21 eb7ac71
Paper minor citation formatting [skip ci]
