### 2019-05-21 #1342 @kgjerde
@leouieda As noted in https://github.com/openjournals/joss/issues/534, I have now fixed the layout issues I wanted to fix in the pdf.
