### 2019-05-21 8f30fd2
Hack correct hyphenation in paper
+ replaced 'highlighted text' with 'pattern matches highlighted'.
