### 2019-05-21 #1342 @kgjerde
Last change in pdf:
* Manually added one case of hyphenation of package name
* replaced 'highlighted text' with 'pattern matches highlighted'.
