### 2019-05-22 9d53829
Fix incomplete argument check
- The "raw" terms variables were not updated when searching, leading argument check to always pass for threshold check.
