+++ inst/explorer/server/2_event_search_button/update_session_variables.R
@@ -36,11 +36,11 @@ search_arguments$case_sensitive <- input$case_sensitivity
 
 # Filtering corpus --------------------------------------------------------
 if (!is.null(input$subset_corpus)) {
-    search_arguments$subset_terms <- collect_subset_terms()
-    search_arguments$subset_thresholds <- collect_threshold_values(search_arguments$subset_terms)
-    search_arguments$subset_custom_column <- collect_custom_column(search_arguments$subset_terms)
+    search_arguments$raw_subset_terms <- collect_subset_terms()  # For later argument check
+    search_arguments$subset_thresholds <- collect_threshold_values(search_arguments$raw_subset_terms)
+    search_arguments$subset_custom_column <- collect_custom_column(search_arguments$raw_subset_terms)
     search_arguments$subset_terms <-
-        to_lower_if_case_insensitive_search(search_arguments$subset_terms) %>%
+        to_lower_if_case_insensitive_search(search_arguments$raw_subset_terms) %>%
         clean_terms()
 }
 
@@ -50,11 +50,11 @@ search_arguments$search_terms <- collect_search_terms() %>%
     unique() %>%
     clean_terms()
 
-#  Highlight terms - collect ----------------------------------------------
-raw_highlight_terms <- collect_highlight_terms()
+#  Highlight terms - collect and keep raw for later argument check---------
+search_arguments$raw_highlight_terms <- collect_highlight_terms()
 
-# Highlight terms - step 1 -------------------------------------------------
-search_arguments$terms_highlight <- raw_highlight_terms %>%
+# Highlight terms - step 1 ------------------------------------------------
+search_arguments$terms_highlight <- search_arguments$raw_highlight_terms %>%
     to_lower_if_case_insensitive_search()
 
 # Identifying duplicates --------------------------------------------------
@@ -66,9 +66,9 @@ search_arguments$terms_highlight <-
     clean_terms()
 
 # Search arguments for all terms (search terms and highlight terms) -------
-search_arguments$thresholds <- collect_threshold_values(raw_highlight_terms) %>%
+search_arguments$thresholds <- collect_threshold_values(search_arguments$raw_highlight_terms) %>%
     .[highlight_terms_not_duplicated]
-search_arguments$custom_column <- collect_custom_column(raw_highlight_terms) %>%
+search_arguments$custom_column <- collect_custom_column(search_arguments$raw_highlight_terms) %>%
     .[highlight_terms_not_duplicated]
 
 # Checking search arguments -----------------------------------------------
