### 2019-05-22 da0afa4
"Reset" subset arguments when new search without subsetting
- Avoids previous search arguments being checked in argument checking
