+++ inst/explorer/server/2_event_search_button/update_session_variables.R
@@ -42,6 +42,13 @@ if (!is.null(input$subset_corpus)) {
     search_arguments$subset_terms <-
         to_lower_if_case_insensitive_search(search_arguments$raw_subset_terms) %>%
         clean_terms()
+} else {
+    # "Resetting" variables if no filtering:
+    search_arguments$subset_search = FALSE
+    search_arguments$subset_terms = NULL
+    search_arguments$subset_thresholds = NA
+    search_arguments$subset_custom_column = NA
+    search_arguments$raw_subset_terms = character(0)
 }
 
 # Search terms ------------------------------------------------------------
