### 2019-05-24 f2ec405
Revamped plot legend
- Produces colours for arbitrary number of search_terms (although some small fixes needed for exceeding six)
- New term starts on new line in legend
- Term with no hits now included in legend with "No hits" text
