+++ inst/explorer/global/corpus_plot_functions/colours_to_plot_and_legend.R
@@ -1,128 +1,271 @@
-# TODO: denne er monstrøs
+# Main function -----------------------------------------------------------
 
 #' Transforming count factors to colours, and legend
 #'
-#' @param test1 Full data frame.
+#' @param main_df Full data frame.
 #' @param linjer Number of search terms.
 #' @param number_of_factors Number of factors for each search term.
+#' @param search_terms_exist The answer to !identical(search_arguments$search_terms, "")
 #'
 #' @return List of length 2: 1) df with column Term_colour for colour of each document/day, 2) df for legend.
-colours_to_plot_and_legend <- function(test1,
+colours_to_plot_and_legend <- function(main_df,
                                        linjer,
-                                       number_of_factors = 8){
+                                       number_of_factors = 8,
+                                       search_terms_exist) {
+  legend_df <- create_legend_df_step_1(main_df)
 
-test1$Term_color <- test1$Term
+  colour_schemes <- define_colours(
+    legend_df = legend_df,
+    linjer = linjer,
+    number_of_factors = number_of_factors
+  )
 
-len_1 <- test1$Term %>%
-    stringr::str_subset("^1") %>%
-    unique %>%
-    length
+  legend_df <-
+    update_legend_df_w_colours(legend_df, colour_schemes)
 
-len_2 <- test1$Term %>%
-    stringr::str_subset("^2") %>%
-    unique %>%
-    length
+  main_df <- populate_df_colour_codes(main_df, legend_df)
 
-len_3 <- test1$Term %>%
-    stringr::str_subset("^3") %>%
-    unique %>%
-    length    
+  legend_df <-
+    add_dummy_colours_to_legend_df(
+      legend_df,
+      linjer,
+      number_of_factors,
+      search_terms_exist
+    )
 
-farger_1 <- character(0)
-farger_2 <- character(0)
-farger_3 <- character(0)
+  main_df <- add_dummy_colours_to_main_df(main_df, legend_df)
+
+  return(list(main_df, legend_df))
+}
+
+# Supporting functions ----------------------------------------------------
 
 # Jeg ønsker spredning av fargene og å unngå de lyseste om mulig.
 # Derfor ulike fargeutvalg avhengig av hvor mange faktorer.
+define_indices_in_colour_palette <-
+  function(number_of_factors) {
+    if (number_of_factors == 8) {
+      fargeutvalg_vektor <- list(
+        9,
+        c(5, 9),
+        c(3, 6, 9),
+        c(3, 5, 7, 9),
+        5:9,
+        4:9,
+        3:9,
+        2:9
+      )
+    } else if (number_of_factors == 4) {
+      fargeutvalg_vektor <- list(
+        4,
+        c(4, 6),
+        4:6,
+        3:6
+      )
+    }
+    return(fargeutvalg_vektor)
+  }
 
-if(number_of_factors == 8){
-fargeutvalg_vektor <- list(
-  9,
-  c(5,9),
-  c(3, 6, 9),
-  c(3, 5, 7, 9),
-  5:9,
-  4:9,
-  3:9,
-  2:9
-)
-} else if(number_of_factors == 4){
-  fargeutvalg_vektor <- list(
-  4,
-  c(4, 6),
-  4:6,
-  3:6
-)
+
+create_legend_df_step_1 <-
+  function(main_df) {
+    legend_df <- tibble::tibble(
+      legend_label = main_df$Term,
+      factor_names = main_df$Term,
+      colour_code = "placeholder"
+    ) %>%
+      dplyr::distinct() %>%
+      dplyr::filter(!is.na(legend_label)) %>%
[…]
