### 2019-05-24 09df557
Allow for arbitrary number of search terms
- in plot functions and in coordinates
