### 2019-05-24 eede4e6
generalise distance between rows to work with arbitrary number of search terms
