### 2019-05-24 0933d89
Improved 365 coordinates when several search terms
- in particular change in day_distance -> increased readability
