### 2019-05-24 2750f4a
Reconnect day corpus mouse click with correct df
-the dynamic [[]] (adjusting to different number of search terms) now covers coordinates for entire tile, encompassing all search terms
