### 2019-05-25 a303a3e
remove redundant argument
