### 2019-05-25 b4d8d67
fix bug from be32c83
- led calendar plot filtered by date to display tiles before and after end of filtering range (from Jan in year 1. Incl Dec in last year)
