### 2019-06-05 #1342 @leouieda
@kgjerde I apologize for letting this hang for so long. Both reviewers have ticked all review boxes. I'll have a final look at the paper first and then we can move on to acceptance.
