### 2019-06-05 #1342 @leouieda
@kgjerde the paper looks good to me so we're ready to move this forward. Before accepting, you'll need to:

1. Release a new version with the changes made after the review (please post the version number here)
2. Archive a copy of source code for this new version on [Zenodo](https://zenodo.org/) and post the DOI for the archive here.

Please don't hesitate to contact me if you have questions or need any help with this.
