### 2019-06-05 78d40ef
Correct capitalisation in references [skip ci]
