### 2019-06-05 #1342 @kgjerde
@leouieda Great!

1. Please note that I now corrected some very few title capitalisation/formatting issues in https://github.com/kgjerde/corporaexplorer/commit/78d40ef and https://github.com/kgjerde/corporaexplorer/commit/284d6d1.

2. I have updated the version number. The version is now 0.5.0.

3. The Zenodo DOI is http://doi.org/10.5281/zenodo.3239097 (see https://zenodo.org/record/3239097).
