### 2019-06-06 2a6ee8e
fix bug introduced by dependency update
- gave error as: Error in : Column `search_term | 5` is of unsupported type function
