### 2019-06-13 #1342 @leouieda
@kgjerde thanks for the updates. I've done a last pass of the paper and we're ready to accept :fireworks:
