### 2019-06-13 #1342 @leouieda
:wave: @openjournals/joss-eics this submission is ready for acceptance
